@@ -1,11 +1,28 @@
 module github.com/muidea/magicOrm
 
-go 1.14
+go 1.23
+
+require github.com/muidea/magicCommon v1.3.122
 
 require (
-	github.com/cihub/seelog v0.0.0-20170130134532-f561c5e57575
-	github.com/go-sql-driver/mysql v1.7.0
-	github.com/leodido/go-urn v1.2.2 // indirect
-	github.com/muidea/magicCommon v0.0.0-20230309140401-af16a770717d
-	golang.org/x/crypto v0.31.0 // indirect
+	filippo.io/edwards25519 v1.1.0 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/gabriel-vasile/mimetype v1.4.6 // indirect
+	github.com/go-playground/locales v0.14.1 // indirect
+	github.com/go-playground/universal-translator v0.18.1 // indirect
+	github.com/go-playground/validator/v10 v10.22.1 // indirect
+	github.com/go-sql-driver/mysql v1.8.1
+	github.com/google/go-cmp v0.7.0
+	github.com/google/uuid v1.6.0 // indirect
+	github.com/leodido/go-urn v1.4.0 // indirect
+	github.com/muidea/seelog v1.0.0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/stretchr/testify v1.10.0
+	golang.org/x/crypto v0.28.0 // indirect
+	golang.org/x/net v0.30.0 // indirect
+	golang.org/x/sys v0.26.0 // indirect
+	golang.org/x/text v0.19.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
+
+replace github.com/muidea/magicCommon => ../magicCommon
