@@ -2,46 +2,96 @@ package mysql
 
 import (
 	"fmt"
+	"strings"
+
+	"github.com/muidea/magicOrm/model"
 )
 
-// EqualOpr EqualOpr
-func EqualOpr(name string, val interface{}) string {
-	return fmt.Sprintf("`%s` = %v", name, val)
-}
+type OprFunc func(string, any, *ResultStack) string
 
-// NotEqualOpr NotEqualOpr
-func NotEqualOpr(name string, val interface{}) string {
-	return fmt.Sprintf("`%s` != %v", name, val)
-}
-
-// BelowOpr BelowOpr
-func BelowOpr(name string, val interface{}) string {
-	return fmt.Sprintf("`%s` < %v", name, val)
-}
-
-// AboveOpr AboveOpr
-func AboveOpr(name string, val interface{}) string {
-	return fmt.Sprintf("`%s` > %v", name, val)
-}
-
-// InOpr InOpr
-func InOpr(name string, val interface{}) string {
-	return fmt.Sprintf("`%s` in (%v)", name, val)
-}
-
-// NotInOpr NotInOpr
-func NotInOpr(name string, val interface{}) string {
-	return fmt.Sprintf("`%s` not in (%v)", name, val)
-}
-
-// LikeOpr LikeOpr
-func LikeOpr(name string, val interface{}) string {
-	valStr, valOK := val.(string)
-	if valOK {
-		return fmt.Sprintf("`%s` LIKE '%%%s%%'", name, valStr[1:len(valStr)-1])
+func getOprFunc(filterItem model.FilterItem) (ret OprFunc) {
+	switch filterItem.OprCode() {
+	case model.EqualOpr:
+		return EqualOpr
+	case model.NotEqualOpr:
+		return NotEqualOpr
+	case model.BelowOpr:
+		return BelowOpr
+	case model.AboveOpr:
+		return AboveOpr
+	case model.InOpr:
+		return InOpr
+	case model.NotInOpr:
+		return NotInOpr
+	case model.LikeOpr:
+		return LikeOpr
 	}
 
-	return ""
+	return nil
+}
+
+// EqualOpr Equal Opr
+func EqualOpr(name string, val any, resultStackPtr *ResultStack) string {
+	resultStackPtr.PushArgs(val)
+	return fmt.Sprintf("`%s` = ?", name)
+}
+
+// NotEqualOpr NotEqual Opr
+func NotEqualOpr(name string, val any, resultStackPtr *ResultStack) string {
+	resultStackPtr.PushArgs(val)
+	return fmt.Sprintf("`%s` != ?", name)
+}
+
+// BelowOpr Below Opr
+func BelowOpr(name string, val any, resultStackPtr *ResultStack) string {
+	resultStackPtr.PushArgs(val)
+	return fmt.Sprintf("`%s` < ?", name)
+}
+
+// AboveOpr Above Opr
+func AboveOpr(name string, val any, resultStackPtr *ResultStack) string {
+	resultStackPtr.PushArgs(val)
+	return fmt.Sprintf("`%s` > ?", name)
+}
+
+// InOpr In Opr
+func InOpr(name string, val any, resultStackPtr *ResultStack) string {
+	sliceVal, sliceOK := val.([]any)
+	if !sliceOK {
+		resultStackPtr.PushArgs(val)
+		return fmt.Sprintf("`%s` in (?)", name)
+	}
+
+	placeHolder := []string{}
+	for _, sv := range sliceVal {
+		resultStackPtr.PushArgs(sv)
+		placeHolder = append(placeHolder, "?")
+	}
+
+	return fmt.Sprintf("`%s` in (%s)", name, strings.Join(placeHolder, ","))
+}
+
+// NotInOpr NotIn Opr
+func NotInOpr(name string, val any, resultStackPtr *ResultStack) string {
+	sliceVal, sliceOK := val.([]any)
+	if !sliceOK {
+		resultStackPtr.PushArgs(val)
+		return fmt.Sprintf("`%s` not in (?)", name)
+	}
+
+	placeHolder := []string{}
+	for _, sv := range sliceVal {
+		resultStackPtr.PushArgs(sv)
+		placeHolder = append(placeHolder, "?")
+	}
+
+	return fmt.Sprintf("`%s` not in (%s)", name, strings.Join(placeHolder, ","))
+}
+
+// LikeOpr Like Opr
+func LikeOpr(name string, val any, resultStackPtr *ResultStack) string {
+	resultStackPtr.PushArgs(fmt.Sprintf("%%%s%%", val))
+	return fmt.Sprintf("`%s` LIKE ?", name)
 }
 
 // SortOpr sort opr
