@@ -2,143 +2,163 @@ package remote
 
 import (
 	"fmt"
-	"reflect"
-	"strings"
 
+	cd "github.com/muidea/magicCommon/def"
+	"github.com/muidea/magicCommon/foundation/log"
 	"github.com/muidea/magicOrm/model"
-	"github.com/muidea/magicOrm/provider/helper"
-	"github.com/muidea/magicOrm/util"
 )
 
-var _helper helper.Helper
-var nilValue model.Value
-
-func init() {
-	_helper = helper.New(ElemDependValue)
-
-	nilValue = &ValueImpl{}
-}
-
-func isRemoteType(vType model.Type) bool {
-	switch vType.GetValue() {
-	case util.TypeDoubleField, util.TypeBooleanField, util.TypeStringField:
-		return true
-	}
-
-	return false
-}
-
-func GetHelper() helper.Helper {
-	return _helper
-}
-
-func GetEntityType(entity interface{}) (ret model.Type, err error) {
-	objPtr, ok := entity.(*Object)
-	if ok {
-		impl := &TypeImpl{Name: objPtr.GetName(), Value: util.TypeStructField, PkgPath: objPtr.GetPkgPath(), IsPtr: objPtr.IsPtr}
-		impl.ElemType = &TypeImpl{Name: objPtr.GetName(), Value: util.TypeStructField, PkgPath: objPtr.GetPkgPath(), IsPtr: objPtr.IsPtr}
-
-		ret = impl
+func GetEntityType(entity any) (ret model.Type, err *cd.Error) {
+	if entity == nil {
+		err = cd.NewError(cd.Unexpected, "entity is nil")
 		return
 	}
 
-	valPtr, ok := entity.(*ObjectValue)
-	if ok {
-		impl := &TypeImpl{Name: valPtr.GetName(), Value: util.TypeStructField, PkgPath: valPtr.GetPkgPath(), IsPtr: valPtr.IsPtr}
-		impl.ElemType = &TypeImpl{Name: valPtr.GetName(), Value: util.TypeStructField, PkgPath: valPtr.GetPkgPath(), IsPtr: valPtr.IsPtr}
-
-		ret = impl
+	switch val := entity.(type) {
+	case *Object:
+		ret = &TypeImpl{
+			Name:    val.Name,
+			PkgPath: val.PkgPath,
+			Value:   model.TypeStructValue,
+			IsPtr:   true,
+		}
+	case *ObjectValue:
+		ret = &TypeImpl{
+			Name:    val.Name,
+			PkgPath: val.PkgPath,
+			Value:   model.TypeStructValue,
+			IsPtr:   true,
+		}
+	case *SliceObjectValue:
+		ret = &TypeImpl{
+			Name:    val.Name,
+			PkgPath: val.PkgPath,
+			Value:   model.TypeSliceValue,
+			IsPtr:   true,
+			ElemType: &TypeImpl{
+				Name:    val.Name,
+				PkgPath: val.PkgPath,
+				Value:   model.TypeStructValue,
+				IsPtr:   true,
+			},
+		}
+	case Object:
+		ret = &TypeImpl{
+			Name:    val.Name,
+			PkgPath: val.PkgPath,
+			Value:   model.TypeStructValue,
+			IsPtr:   true,
+		}
+	case ObjectValue:
+		ret = &TypeImpl{
+			Name:    val.Name,
+			PkgPath: val.PkgPath,
+			Value:   model.TypeStructValue,
+			IsPtr:   true,
+		}
+	case SliceObjectValue:
+		ret = &TypeImpl{
+			Name:    val.Name,
+			PkgPath: val.PkgPath,
+			Value:   model.TypeSliceValue,
+			IsPtr:   true,
+			ElemType: &TypeImpl{
+				Name:    val.Name,
+				PkgPath: val.PkgPath,
+				Value:   model.TypeStructValue,
+				IsPtr:   true,
+			},
+		}
+	default:
+		err = cd.NewError(cd.Unexpected, fmt.Sprintf("illegal entity, entity:%v", entity))
 		return
 	}
-
-	sValPtr, ok := entity.(*SliceObjectValue)
-	if ok {
-		impl := &TypeImpl{Name: sValPtr.GetName(), Value: util.TypeSliceField, PkgPath: sValPtr.GetPkgPath(), IsPtr: sValPtr.IsPtr}
-		impl.ElemType = &TypeImpl{Name: sValPtr.GetName(), Value: util.TypeStructField, PkgPath: sValPtr.GetPkgPath(), IsPtr: sValPtr.IsElemPtr}
-
-		ret = impl
-		return
-	}
-
-	typeImpl, typeErr := newType(reflect.TypeOf(entity))
-	if typeErr != nil {
-		err = typeErr
-		return
-	}
-	if !isRemoteType(typeImpl.Elem()) {
-		err = fmt.Errorf("illegal entity type, name:%s", typeImpl.GetName())
-		return
-	}
-
-	ret = typeImpl
 	return
 }
 
-func GetEntityValue(entity interface{}) (ret model.Value, err error) {
-	rVal := reflect.ValueOf(entity)
-	ret = newValue(rVal)
+func GetEntityValue(entity any) (ret model.Value, err *cd.Error) {
+	if entity == nil {
+		err = cd.NewError(cd.Unexpected, "entity is nil")
+		return
+	}
+
+	switch val := entity.(type) {
+	case *ObjectValue:
+		ret = &ValueImpl{
+			value: val,
+		}
+	case *SliceObjectValue:
+		ret = &ValueImpl{
+			value: val,
+		}
+	case ObjectValue:
+		ret = &ValueImpl{
+			value: &val,
+		}
+	case SliceObjectValue:
+		ret = &ValueImpl{
+			value: &val,
+		}
+	default:
+		err = cd.NewError(cd.Unexpected, fmt.Sprintf("illegal entity, entity:%v", entity))
+		//log.Errorf("GetEntityValue failed, err:%s", err.Error())
+		return
+	}
+
 	return
 }
 
-func GetEntityModel(entity interface{}) (ret model.Model, err error) {
-	objPtr, ok := entity.(*Object)
-	if ok {
-		ret = objPtr
+func GetEntityModel(entity any) (ret model.Model, err *cd.Error) {
+	if entity == nil {
+		err = cd.NewError(cd.Unexpected, "entity is nil")
 		return
 	}
 
-	err = fmt.Errorf("illegal value, not object")
+	switch val := entity.(type) {
+	case *Object:
+		ret = val
+	case Object:
+		ret = &val
+	default:
+		err = cd.NewError(cd.Unexpected, fmt.Sprintf("illegal entity, entity:%v", entity))
+		log.Errorf("GetEntityModel failed, err:%s", err.Error())
+	}
+
 	return
 }
 
-func SetModelValue(vModel model.Model, vVal model.Value) (ret model.Model, err error) {
-	rVal := vVal.Get()
-	nameVal := rVal.FieldByName("Name")
-	pkgVal := rVal.FieldByName("PkgPath")
-	itemsVal := rVal.FieldByName("Items")
-	if util.IsNil(nameVal) || util.IsNil(pkgVal) || util.IsNil(itemsVal) {
-		err = fmt.Errorf("illegal model value")
-		return
-	}
-	if nameVal.String() != vModel.GetName() || pkgVal.String() != vModel.GetPkgPath() {
-		err = fmt.Errorf("illegal model value, mismatch model value")
+func GetModelFilter(vModel model.Model) (ret model.Filter, err *cd.Error) {
+	objectPtr, objectOK := vModel.(*Object)
+	if !objectOK {
+		err = cd.NewError(cd.Unexpected, fmt.Sprintf("illegal model, model:%v", vModel))
+		log.Errorf("GetModelFilter failed, err:%s", err.Error())
 		return
 	}
 
-	for idx := 0; idx < itemsVal.Len(); idx++ {
-		iVal := reflect.Indirect(itemsVal.Index(idx))
+	ret = NewFilter(objectPtr)
+	return
+}
 
-		iName := iVal.FieldByName("Name").String()
-		iValue := iVal.FieldByName("Value")
-		if iValue.Kind() == reflect.Interface {
-			iValue = iValue.Elem()
+func SetModelValue(vModel model.Model, vVal model.Value) (ret model.Model, err *cd.Error) {
+	defer func() {
+		if errInfo := recover(); errInfo != nil {
+			err = cd.NewError(cd.Unexpected, fmt.Sprintf("SetModelValue failed, illegal value, err:%v", errInfo))
+			log.Errorf("SetModelValue failed, err:%s", err.Error())
+			return
 		}
+	}()
 
-		vField := vModel.GetField(iName)
-		if util.IsNil(iValue) {
-			vField.SetValue(nilValue)
-			continue
+	switch val := vVal.Get().(type) {
+	case *ObjectValue:
+		err = assignObjectValue(vModel, val)
+	default:
+		if vVal.IsValid() {
+			err = vModel.SetPrimaryFieldValue(val)
+		} else {
+			err = cd.NewError(cd.Unexpected, fmt.Sprintf("illegal model value, val:%v", val))
 		}
-
-		vType := vField.GetType()
-		if vType.IsBasic() {
-			vValue, vErr := _helper.Decode(iValue.Interface(), vField.GetType())
-			if vErr != nil {
-				err = vErr
-				return
-			}
-
-			err = vField.SetValue(vValue)
-			if err != nil {
-				return
-			}
-
-			continue
-		}
-
-		vValue := newValue(iValue)
-		err = vField.SetValue(vValue)
 		if err != nil {
+			log.Errorf("SetModelValue failed, err:%s", err.Error())
 			return
 		}
 	}
@@ -147,149 +167,15 @@ func SetModelValue(vModel model.Model, vVal model.Value) (ret model.Model, err e
 	return
 }
 
-func ElemDependValue(vVal model.Value) (ret []model.Value, err error) {
-	rVal := vVal.Get()
-	if rVal.Type().String() == reflect.TypeOf(_declareObjectSliceValue).String() {
-		objectsVal := rVal.FieldByName("Values")
-		if !util.IsNil(objectsVal) {
-			for idx := 0; idx < objectsVal.Len(); idx++ {
-				ret = append(ret, newValue(objectsVal.Index(idx)))
-			}
-
+func assignObjectValue(vModel model.Model, objectValuePtr *ObjectValue) (err *cd.Error) {
+	for idx := range objectValuePtr.Fields {
+		fieldVal := objectValuePtr.Fields[idx]
+		err = vModel.SetFieldValue(fieldVal.GetName(), fieldVal.Get())
+		if err != nil {
+			log.Errorf("assignObjectValue failed, err:%s", err.Error())
 			return
 		}
 	}
 
-	if rVal.Type().String() == reflect.TypeOf(_declareObjectValue).String() {
-		itemsVal := rVal.FieldByName("Items")
-		if !util.IsNil(itemsVal) {
-			ret = append(ret, vVal)
-			return
-		}
-	}
-
-	tVal, tErr := util.GetTypeEnum(rVal.Type())
-	if tErr != nil {
-		err = tErr
-		return
-	}
-
-	if util.IsSliceType(tVal) {
-		for idx := 0; idx < rVal.Len(); idx++ {
-			ret = append(ret, newValue(rVal.Index(idx)))
-		}
-		return
-	}
-
-	if util.IsBasicType(tVal) {
-		ret = append(ret, vVal)
-		return
-	}
-
-	err = fmt.Errorf("illegal remote slice value, type:%s", rVal.Type().String())
-	return
-}
-
-func AppendSliceValue(sliceVal model.Value, vVal model.Value) (ret model.Value, err error) {
-	sliceName := sliceVal.Get().FieldByName("Name").String()
-	slicePkg := sliceVal.Get().FieldByName("PkgPath").String()
-	objectName := vVal.Get().FieldByName("Name").String()
-	objectPkg := vVal.Get().FieldByName("PkgPath").String()
-
-	if sliceName != objectName || slicePkg != objectPkg {
-		err = fmt.Errorf("mismatch slice value")
-		return
-	}
-
-	sliceObjects := sliceVal.Get().FieldByName("Values")
-	if util.IsNil(sliceObjects) {
-		err = fmt.Errorf("illegal remote model slice value")
-		return
-	}
-
-	sliceObjects = reflect.Append(sliceObjects, vVal.Get().Addr())
-	sliceVal.Get().FieldByName("Values").Set(sliceObjects)
-
-	ret = sliceVal
-	return
-}
-
-func encodeModel(vVal model.Value, vType model.Type, mCache model.Cache, helper helper.Helper) (ret interface{}, err error) {
-	tModel := mCache.Fetch(vType.GetName())
-	if tModel == nil {
-		err = fmt.Errorf("illegal value type,type:%s", vType.GetName())
-		return
-	}
-
-	if vVal.IsBasic() {
-		pkField := tModel.GetPrimaryField()
-		vType := pkField.GetType()
-		ret, err = helper.Encode(vVal, vType)
-		return
-	}
-
-	vModel, vErr := SetModelValue(tModel.Copy(), vVal)
-	if vErr != nil {
-		err = vErr
-		return
-	}
-
-	pkField := vModel.GetPrimaryField()
-	tType := pkField.GetType()
-	tVal := pkField.GetValue()
-	if tVal.IsNil() {
-		tVal, _ = tType.Interface()
-	}
-
-	ret, err = helper.Encode(tVal, tType)
-	return
-}
-
-func encodeSliceModel(tVal model.Value, tType model.Type, mCache model.Cache, helper helper.Helper) (ret string, err error) {
-	vVals, vErr := ElemDependValue(tVal)
-	if vErr != nil {
-		err = vErr
-		return
-	}
-	if len(vVals) == 0 {
-		return
-	}
-
-	items := []string{}
-	for _, v := range vVals {
-		strVal, strErr := encodeModel(v, tType.Elem(), mCache, helper)
-		if strErr != nil {
-			err = strErr
-			return
-		}
-
-		items = append(items, fmt.Sprintf("%v", strVal))
-	}
-
-	ret = strings.Join(items, ",")
-	return
-}
-
-func EncodeValue(tVal model.Value, tType model.Type, mCache model.Cache) (ret interface{}, err error) {
-	if tType.IsBasic() {
-		ret, err = _helper.Encode(tVal, tType)
-		return
-	}
-	if util.IsStructType(tType.GetValue()) {
-		ret, err = encodeModel(tVal, tType, mCache, _helper)
-		return
-	}
-
-	ret, err = encodeSliceModel(tVal, tType, mCache, _helper)
-	return
-}
-
-func DecodeValue(tVal interface{}, tType model.Type, mCache model.Cache) (ret model.Value, err error) {
-	if tType.IsBasic() {
-		ret, err = _helper.Decode(tVal, tType)
-		return
-	}
-
-	err = fmt.Errorf("unexecption type, type name:%s", tType.GetName())
 	return
 }
