@@ -1,36 +1,142 @@
+//go:build local || all
+// +build local all
+
 package test
 
 import (
-	"github.com/muidea/magicOrm/orm"
-	"github.com/muidea/magicOrm/provider"
 	"testing"
 	"time"
+
+	"github.com/muidea/magicCommon/foundation/util"
+	"github.com/muidea/magicOrm/orm"
+	"github.com/muidea/magicOrm/provider"
 )
 
 const localOwner = "local"
 
-func TestLocalSimple(t *testing.T) {
+func TestOptional(t *testing.T) {
 	orm.Initialize()
-	defer orm.Uninitialize()
+	defer orm.Uninitialized()
 
-	config := orm.NewConfig("localhost:3306", "testdb", "root", "rootkit")
+	config := orm.NewConfig("localhost:3306", "testdb", "root", "rootkit", "")
 	localProvider := provider.NewLocalProvider(localOwner)
 
-	o1, err := orm.NewOrm(localProvider, config)
+	o1, err := orm.NewOrm(localProvider, config, "abc")
 	defer o1.Release()
 	if err != nil {
 		t.Errorf("new Orm failed, err:%s", err.Error())
 		return
 	}
 
-	objList := []interface{}{&Simple{}}
-	_, err = registerModel(localProvider, objList)
+	objList := []any{&Optional{}}
+	_, err = registerLocalModel(localProvider, objList)
 	if err != nil {
 		t.Errorf("register model failed. err:%s", err.Error())
 		return
 	}
 
-	ts, _ := time.Parse("2006-01-02 15:04:05", "2018-01-02 15:04:05")
+	opt001 := Optional{Name: "abc", StrArry: []string{"a", "b", "c"}}
+	optionalModel, err := localProvider.GetEntityModel(opt001)
+	if err != nil {
+		t.Errorf("GetEntityModel failed, err:%s", err.Error())
+		return
+	}
+
+	err = o1.Drop(optionalModel)
+	if err != nil {
+		t.Errorf("drop optional failed, err:%s", err.Error())
+		return
+	}
+
+	err = o1.Create(optionalModel)
+	if err != nil {
+		t.Errorf("create optional failed, err:%s", err.Error())
+		return
+	}
+
+	newOpt000 := optionalModel.Interface(true).(*Optional)
+	if newOpt000.Name != opt001.Name {
+		t.Errorf("insert optional failed, missmatch name, expect:%s, actual:%s", opt001.Name, newOpt000.Name)
+	}
+	if newOpt000.Optional != nil {
+		t.Errorf("insert optional failed, missmatch optional, expect nil, actual:%+v", newOpt000.Optional)
+	}
+	if newOpt000.OptionnalStrArray != nil {
+		t.Errorf("insert optional failed, missmatch optionnalStrArray, expect nil, actual:%+v", newOpt000.OptionnalStrArray)
+	}
+
+	optionalModel, err = o1.Insert(optionalModel)
+	if err != nil {
+		t.Errorf("insert optional failed, err:%s", err.Error())
+	}
+	newOpt001 := optionalModel.Interface(true).(*Optional)
+	if newOpt001.Name != opt001.Name {
+		t.Errorf("insert optional failed, missmatch name, expect:%s, actual:%s", opt001.Name, newOpt001.Name)
+	}
+	if newOpt001.Optional != nil {
+		t.Errorf("insert optional failed, missmatch optional, expect nil, actual:%+v", newOpt001.Optional)
+	}
+	if newOpt001.OptionnalStrArray != nil {
+		t.Errorf("insert optional failed, missmatch optionnalStrArray, expect nil, actual:%+v", newOpt001.OptionnalStrArray)
+	}
+
+	err = optionalModel.SetFieldValue("name", "def")
+	if err != nil {
+		t.Errorf("set optional failed, err:%s", err.Error())
+	}
+	optionalModel, err = o1.Update(optionalModel)
+	if err != nil {
+		t.Errorf("update optional failed, err:%s", err.Error())
+	}
+	newOpt002 := optionalModel.Interface(true).(*Optional)
+	if newOpt002.Name != "def" {
+		t.Errorf("update optional failed, missmatch name, expect:%s, actual:%s", "def", newOpt002.Name)
+	}
+
+	err = optionalModel.SetFieldValue("name", "ghi")
+	if err != nil {
+		t.Errorf("set optional failed, err:%s", err.Error())
+	}
+	newPtr := &opt001.Name
+	err = optionalModel.SetFieldValue("optional", newPtr)
+	if err != nil {
+		t.Errorf("set optional failed, err:%s", err.Error())
+	}
+	optionalModel, err = o1.Update(optionalModel)
+	if err != nil {
+		t.Errorf("update optional failed, err:%s", err.Error())
+	}
+	newOpt003 := optionalModel.Interface(true).(*Optional)
+	if newOpt003.Name != "ghi" {
+		t.Errorf("update optional failed, missmatch name, expect:%s, actual:%s", "ghi", newOpt003.Name)
+	}
+	if *newOpt003.Optional != opt001.Name {
+		t.Errorf("update optional failed, missmatch optional, expect:%s, actual:%s", opt001.Name, *newOpt003.Optional)
+	}
+}
+
+func TestLocalSimple(t *testing.T) {
+	orm.Initialize()
+	defer orm.Uninitialized()
+
+	config := orm.NewConfig("localhost:3306", "testdb", "root", "rootkit", "")
+	localProvider := provider.NewLocalProvider(localOwner)
+
+	o1, err := orm.NewOrm(localProvider, config, "abc")
+	defer o1.Release()
+	if err != nil {
+		t.Errorf("new Orm failed, err:%s", err.Error())
+		return
+	}
+
+	objList := []any{&Simple{}}
+	_, err = registerLocalModel(localProvider, objList)
+	if err != nil {
+		t.Errorf("register model failed. err:%s", err.Error())
+		return
+	}
+
+	ts, _ := time.Parse(util.CSTLayout, "2018-01-02 15:04:05")
 	s1 := &Simple{I8: 12, I16: 23, I32: 34, I64: 45, Name: "test code", Value: 12.345, F64: 23.456, TimeStamp: ts, Flag: true}
 
 	s1Model, s1Err := localProvider.GetEntityModel(s1)
@@ -72,7 +178,7 @@ func TestLocalSimple(t *testing.T) {
 	s1 = s1Model.Interface(true).(*Simple)
 
 	s2 := Simple{ID: s1.ID}
-	s2Model, s2Err := localProvider.GetEntityModel(s2)
+	s2Model, s2Err := localProvider.GetEntityModel(&s2)
 	if s2Err != nil {
 		t.Errorf("GetEntityModel failed, err:%s", s2Err.Error())
 		return
@@ -92,26 +198,26 @@ func TestLocalSimple(t *testing.T) {
 
 func TestLocalReference(t *testing.T) {
 	orm.Initialize()
-	defer orm.Uninitialize()
+	defer orm.Uninitialized()
 
-	config := orm.NewConfig("localhost:3306", "testdb", "root", "rootkit")
+	config := orm.NewConfig("localhost:3306", "testdb", "root", "rootkit", "")
 	localProvider := provider.NewLocalProvider(localOwner)
 
-	o1, err := orm.NewOrm(localProvider, config)
+	o1, err := orm.NewOrm(localProvider, config, "abc")
 	defer o1.Release()
 	if err != nil {
 		t.Errorf("new Orm failed, err:%s", err.Error())
 		return
 	}
 
-	objList := []interface{}{&Reference{}}
-	_, err = registerModel(localProvider, objList)
+	objList := []any{&Reference{}}
+	_, err = registerLocalModel(localProvider, objList)
 	if err != nil {
 		t.Errorf("register model failed. err:%s", err.Error())
 		return
 	}
 
-	ts, _ := time.Parse("2006-01-02 15:04:05", "2018-01-02 15:04:05")
+	ts, _ := time.Parse(util.CSTLayout, "2018-01-02 15:04:05")
 	strValue := "test code"
 	fValue := float32(12.34)
 	flag := true
@@ -119,13 +225,13 @@ func TestLocalReference(t *testing.T) {
 	fArray := []float32{12.34, 23, 45, 45, 67}
 	strArray := []string{"Abc", "Bcd"}
 	bArray := []bool{true, true, false, false}
-	strPtrArray := []*string{&strValue, &strValue}
+	strPtrArray := []string{strValue, strValue}
 	s1 := &Reference{
 		Name:        strValue,
-		FValue:      &fValue,
+		FValue:      fValue,
 		F64:         23.456,
-		TimeStamp:   &ts,
-		Flag:        &flag,
+		TimeStamp:   ts,
+		Flag:        flag,
 		IArray:      iArray,
 		FArray:      fArray,
 		StrArray:    strArray,
@@ -176,18 +282,23 @@ func TestLocalReference(t *testing.T) {
 	fValue2 := float32(0.0)
 	var ts2 time.Time
 	var bVal bool
-	var strArray2 []string
-	var ptrArray2 []*string
+	strArray2 := []string{}
+	ptrArray2 := []string{}
 	s2 := Reference{
 		ID:          s1.ID,
-		FValue:      &fValue2,
-		TimeStamp:   &ts2,
-		Flag:        &bVal,
+		FValue:      fValue2,
+		TimeStamp:   ts2,
+		Flag:        bVal,
+		IArray:      []int{},
+		FArray:      []float32{},
+		StrArray:    []string{},
+		BArray:      []bool{},
 		PtrArray:    &strArray2,
+		StrPtrArray: []string{},
 		PtrStrArray: &ptrArray2,
 	}
 
-	s2Model, s2Err := localProvider.GetEntityModel(s2)
+	s2Model, s2Err := localProvider.GetEntityModel(&s2)
 	if s2Err != nil {
 		t.Errorf("GetEntityModel failed, err:%s", s2Err.Error())
 		return
@@ -216,10 +327,12 @@ func TestLocalReference(t *testing.T) {
 		return
 	}
 
+	newPtrArray := []string{}
 	s4 := Reference{
-		ID: s1.ID,
+		ID:       s1.ID,
+		PtrArray: &newPtrArray,
 	}
-	s4Model, s4Err := localProvider.GetEntityModel(s4)
+	s4Model, s4Err := localProvider.GetEntityModel(&s4)
 	if s4Err != nil {
 		t.Errorf("GetEntityModel failed, err:%s", s4Err.Error())
 		return
@@ -235,28 +348,28 @@ func TestLocalReference(t *testing.T) {
 		t.Errorf("query reference failed, err:%s", err.Error())
 		return
 	}
-	if s4.FValue != nil || s4.TimeStamp != nil || s4.Flag != nil || s4.PtrStrArray != nil || s4.PtrArray != nil {
-		t.Errorf("query reference failed, err:%s", err.Error())
+	if s4.IArray == nil || s4.FArray == nil || s4.PtrStrArray != nil || s4.PtrArray == nil {
+		t.Errorf("query reference failed")
 		return
 	}
 }
 
 func TestLocalCompose(t *testing.T) {
 	orm.Initialize()
-	defer orm.Uninitialize()
+	defer orm.Uninitialized()
 
-	config := orm.NewConfig("localhost:3306", "testdb", "root", "rootkit")
+	config := orm.NewConfig("localhost:3306", "testdb", "root", "rootkit", "")
 	localProvider := provider.NewLocalProvider(localOwner)
 
-	o1, err := orm.NewOrm(localProvider, config)
+	o1, err := orm.NewOrm(localProvider, config, "abc")
 	defer o1.Release()
 	if err != nil {
 		t.Errorf("new Orm failed, err:%s", err.Error())
 		return
 	}
 
-	objList := []interface{}{&Simple{}, &Reference{}, &Compose{}}
-	mList, mErr := registerModel(localProvider, objList)
+	objList := []any{&Simple{}, &Reference{}, &Compose{}}
+	mList, mErr := registerLocalModel(localProvider, objList)
 	if mErr != nil {
 		err = mErr
 		t.Errorf("register model failed. err:%s", err.Error())
@@ -277,7 +390,7 @@ func TestLocalCompose(t *testing.T) {
 		}
 	}
 
-	ts, _ := time.Parse("2006-01-02 15:04:05", "2018-01-02 15:04:05")
+	ts, _ := time.Parse(util.CSTLayout, "2018-01-02 15:04:05")
 	s1 := Simple{I8: 12, I16: 23, I32: 34, I64: 45, Name: "test code", Value: 12.345, F64: 23.456, TimeStamp: ts, Flag: true}
 	s1Model, s1Err := localProvider.GetEntityModel(s1)
 	if s1Err != nil {
@@ -298,13 +411,13 @@ func TestLocalCompose(t *testing.T) {
 	fArray := []float32{12.34, 23, 45, 45, 67}
 	strArray := []string{"Abc", "Bcd"}
 	bArray := []bool{true, true, false, false}
-	strPtrArray := []*string{&strValue, &strValue}
+	strPtrArray := []string{strValue, strValue}
 	r1 := Reference{
 		Name:        strValue,
-		FValue:      &fValue,
+		FValue:      fValue,
 		F64:         23.456,
-		TimeStamp:   &ts,
-		Flag:        &flag,
+		TimeStamp:   ts,
+		Flag:        flag,
 		IArray:      iArray,
 		FArray:      fArray,
 		StrArray:    strArray,
@@ -314,7 +427,7 @@ func TestLocalCompose(t *testing.T) {
 		PtrStrArray: &strPtrArray,
 	}
 
-	r1Model, r1Err := localProvider.GetEntityModel(r1)
+	r1Model, r1Err := localProvider.GetEntityModel(&r1)
 	if r1Err != nil {
 		t.Errorf("GetEntityModel failed, err:%s", r1Err.Error())
 		return
@@ -330,16 +443,15 @@ func TestLocalCompose(t *testing.T) {
 
 	refPtrArray := []*Reference{&r1}
 	c1 := &Compose{
-		Name:           strValue,
-		Simple:         s1,
-		PtrSimple:      &s1,
-		SimpleArray:    []Simple{s1, s1},
-		SimplePtrArray: []*Simple{&s1, &s1},
-		Reference:      r1,
-		PtrReference:   &r1,
-		RefArray:       []Reference{r1, r1, r1},
-		RefPtrArray:    refPtrArray,
-		PtrRefArray:    &refPtrArray,
+		Name:              strValue,
+		Simple:            s1,
+		SimplePtr:         &s1,
+		SimpleArray:       []Simple{s1, s1},
+		SimplePtrArray:    []*Simple{&s1, &s1},
+		Reference:         r1,
+		ReferencePtr:      &r1,
+		ReferenceArray:    []Reference{r1, r1, r1},
+		ReferencePtrArray: refPtrArray,
 	}
 	c1Model, c1Err := localProvider.GetEntityModel(c1)
 	if c1Err != nil {
@@ -354,19 +466,18 @@ func TestLocalCompose(t *testing.T) {
 	c1 = c1Model.Interface(true).(*Compose)
 
 	c2 := Compose{
-		Name:           strValue,
-		Simple:         s1,
-		PtrSimple:      &s1,
-		SimpleArray:    []Simple{s1, s1},
-		SimplePtrArray: []*Simple{&s1, &s1},
-		Reference:      r1,
-		PtrReference:   &r1,
-		RefArray:       []Reference{r1, r1, r1},
-		RefPtrArray:    refPtrArray,
-		PtrRefArray:    &refPtrArray,
-		PtrCompose:     c1,
+		Name:              strValue,
+		Simple:            s1,
+		SimplePtr:         &s1,
+		SimpleArray:       []Simple{s1, s1},
+		SimplePtrArray:    []*Simple{&s1, &s1},
+		Reference:         r1,
+		ReferencePtr:      &r1,
+		ReferenceArray:    []Reference{r1, r1, r1},
+		ReferencePtrArray: refPtrArray,
+		ComposePtr:        c1,
 	}
-	c2Model, c2Err := localProvider.GetEntityModel(c2)
+	c2Model, c2Err := localProvider.GetEntityModel(&c2)
 	if c2Err != nil {
 		t.Errorf("GetEntityModel failed,err:%s", c2Err.Error())
 		return
@@ -380,17 +491,17 @@ func TestLocalCompose(t *testing.T) {
 	c2 = c2Model.Interface(false).(Compose)
 
 	c3 := Compose{
-		ID:             c2.ID,
-		PtrSimple:      &Simple{},
-		SimpleArray:    []Simple{},
-		SimplePtrArray: []*Simple{},
-		PtrSimpleArray: &[]Simple{},
-		PtrReference:   &Reference{},
-		RefArray:       []Reference{},
-		PtrRefArray:    &[]*Reference{},
-		PtrCompose:     &Compose{},
+		ID:                c2.ID,
+		SimplePtr:         &Simple{},
+		SimpleArray:       []Simple{},
+		SimplePtrArray:    []*Simple{},
+		SimpleArrayPtr:    &[]Simple{},
+		ReferencePtr:      &Reference{},
+		ReferenceArray:    []Reference{},
+		ReferencePtrArray: []*Reference{},
+		ComposePtr:        &Compose{},
 	}
-	c3Model, c3Err := localProvider.GetEntityModel(c3)
+	c3Model, c3Err := localProvider.GetEntityModel(&c3)
 	if c3Err != nil {
 		t.Errorf("GetEntityModel failed, err:%s", c3Err.Error())
 		return
@@ -415,20 +526,20 @@ func TestLocalCompose(t *testing.T) {
 
 func TestLocalQuery(t *testing.T) {
 	orm.Initialize()
-	defer orm.Uninitialize()
+	defer orm.Uninitialized()
 
-	config := orm.NewConfig("localhost:3306", "testdb", "root", "rootkit")
+	config := orm.NewConfig("localhost:3306", "testdb", "root", "rootkit", "")
 	localProvider := provider.NewLocalProvider(localOwner)
 
-	o1, err := orm.NewOrm(localProvider, config)
+	o1, err := orm.NewOrm(localProvider, config, "abc")
 	defer o1.Release()
 	if err != nil {
 		t.Errorf("new Orm failed, err:%s", err.Error())
 		return
 	}
 
-	objList := []interface{}{&Simple{}, &Reference{}, &Compose{}}
-	mList, mErr := registerModel(localProvider, objList)
+	objList := []any{&Simple{}, &Reference{}, &Compose{}}
+	mList, mErr := registerLocalModel(localProvider, objList)
 	if mErr != nil {
 		t.Errorf("register model failed. err:%s", mErr.Error())
 		return
@@ -448,7 +559,7 @@ func TestLocalQuery(t *testing.T) {
 		}
 	}
 
-	ts, _ := time.ParseInLocation("2006-01-02 15:04:05", "2018-01-02 15:04:05", time.Local)
+	ts, _ := time.ParseInLocation(util.CSTLayout, "2018-01-02 15:04:05", time.Local)
 	s1 := Simple{I8: 12, I16: 23, I32: 34, I64: 45, Name: "test code", Value: 12.345, F64: 23.456, TimeStamp: ts, Flag: true}
 	s1Model, s1Err := localProvider.GetEntityModel(s1)
 	if s1Err != nil {
@@ -470,13 +581,13 @@ func TestLocalQuery(t *testing.T) {
 	fArray := []float32{12.34, 23, 45, 45, 67}
 	strArray := []string{"Abc", "Bcd"}
 	bArray := []bool{true, true, false, false}
-	strPtrArray := []*string{&strValue, &strValue}
+	strPtrArray := []string{strValue, strValue}
 	r1 := Reference{
 		Name:        strValue,
-		FValue:      &fValue,
+		FValue:      fValue,
 		F64:         23.456,
-		TimeStamp:   &ts,
-		Flag:        &flag,
+		TimeStamp:   ts,
+		Flag:        flag,
 		IArray:      iArray,
 		FArray:      fArray,
 		StrArray:    strArray,
@@ -499,16 +610,15 @@ func TestLocalQuery(t *testing.T) {
 
 	refPtrArray := []*Reference{&r1}
 	c1 := Compose{
-		Name:           strValue,
-		Simple:         s1,
-		PtrSimple:      &s1,
-		SimpleArray:    []Simple{s1, s1},
-		SimplePtrArray: []*Simple{&s1, &s1},
-		Reference:      r1,
-		PtrReference:   &r1,
-		RefArray:       []Reference{r1, r1, r1},
-		RefPtrArray:    refPtrArray,
-		PtrRefArray:    &refPtrArray,
+		Name:              strValue,
+		Simple:            s1,
+		SimplePtr:         &s1,
+		SimpleArray:       []Simple{s1, s1},
+		SimplePtrArray:    []*Simple{&s1, &s1},
+		Reference:         r1,
+		ReferencePtr:      &r1,
+		ReferenceArray:    []Reference{r1, r1, r1},
+		ReferencePtrArray: refPtrArray,
 	}
 	c1Model, c1Err := localProvider.GetEntityModel(c1)
 	if c1Err != nil {
@@ -525,17 +635,16 @@ func TestLocalQuery(t *testing.T) {
 
 	strValue = "123"
 	c2 := Compose{
-		Name:           strValue,
-		Simple:         s1,
-		PtrSimple:      &s1,
-		SimpleArray:    []Simple{s1, s1},
-		SimplePtrArray: []*Simple{&s1, &s1},
-		Reference:      r1,
-		PtrReference:   &r1,
-		RefArray:       []Reference{r1, r1, r1},
-		RefPtrArray:    refPtrArray,
-		PtrRefArray:    &refPtrArray,
-		PtrCompose:     &c1,
+		Name:              strValue,
+		Simple:            s1,
+		SimplePtr:         &s1,
+		SimpleArray:       []Simple{s1, s1},
+		SimplePtrArray:    []*Simple{&s1, &s1},
+		Reference:         r1,
+		ReferencePtr:      &r1,
+		ReferenceArray:    []Reference{r1, r1, r1},
+		ReferencePtrArray: refPtrArray,
+		ComposePtr:        &c1,
 	}
 	c2Model, c2Err := localProvider.GetEntityModel(c2)
 	if c2Err != nil {
@@ -561,7 +670,7 @@ func TestLocalQuery(t *testing.T) {
 		t.Errorf("insert compose failed, err:%s", err.Error())
 		return
 	}
-	c3 = c3Model.Interface(false).(Compose)
+	_ = c3Model.Interface(false).(Compose)
 
 	c4 := c2
 	c4Model, c4Err := localProvider.GetEntityModel(c4)
@@ -574,15 +683,15 @@ func TestLocalQuery(t *testing.T) {
 		t.Errorf("insert compose failed, err:%s", err.Error())
 		return
 	}
-	c4 = c4Model.Interface(false).(Compose)
+	_ = c4Model.Interface(false).(Compose)
 
-	cList := []*Compose{}
-	cModel, cErr := localProvider.GetEntityModel(cList)
-	if cErr != nil {
-		t.Errorf("GetEntityModel failed, err:%s", cErr.Error())
+	cModel, _ := localProvider.GetEntityModel(&Compose{})
+	filter, err := localProvider.GetModelFilter(cModel)
+	if err != nil {
+		t.Errorf("GetEntityFilter failed, err:%s", err.Error())
 		return
 	}
-	cModelList, cModelErr := o1.BatchQuery(cModel, nil)
+	cModelList, cModelErr := o1.BatchQuery(filter)
 	if cModelErr != nil {
 		t.Errorf("batch query compose failed, err:%s", cModelErr.Error())
 		return
@@ -592,17 +701,115 @@ func TestLocalQuery(t *testing.T) {
 		return
 	}
 
-	cList = []*Compose{}
-	filter := orm.GetFilter(localProvider)
 	filter.Equal("name", c2.Name)
-	filter.ValueMask(&Compose{PtrSimple: &Simple{}})
-	cModelList, cModelErr = o1.BatchQuery(cModel, filter)
+	filter.ValueMask(&Compose{SimplePtr: &Simple{}})
+	cModelList, cModelErr = o1.BatchQuery(filter)
 	if cModelErr != nil {
 		t.Errorf("batch query compose failed, err:%s", cModelErr.Error())
 		return
 	}
 	if len(cModelList) != 3 {
 		t.Errorf("batch query compose failed")
+		return
+	}
+}
+
+func TestLocalOnlineEntity(t *testing.T) {
+	orm.Initialize()
+	defer orm.Uninitialized()
+
+	config := orm.NewConfig("localhost:3306", "testdb", "root", "rootkit", "")
+	localProvider := provider.NewLocalProvider(localOwner)
+
+	o1, err := orm.NewOrm(localProvider, config, "abc")
+	defer o1.Release()
+	if err != nil {
+		t.Errorf("new Orm failed, err:%s", err.Error())
+		return
+	}
+
+	objList := []any{&Entity{}, &OnlineEntity{}}
+	mList, mErr := registerLocalModel(localProvider, objList)
+	if mErr != nil {
+		t.Errorf("register model failed. err:%s", mErr.Error())
+		return
+	}
+
+	err = dropModel(o1, mList)
+	if err != nil {
+		t.Errorf("drop model failed. err:%s", err.Error())
+		return
+	}
+
+	err = createModel(o1, mList)
+	if err != nil {
+		t.Errorf("create model failed. err:%s", err.Error())
+		return
+	}
+
+	entityPtr := &Entity{EName: "test", EType: "test", EID: 10, Namespace: "test"}
+	entityModelVal, entityModelErr := localProvider.GetEntityModel(entityPtr)
+	if entityModelErr != nil {
+		t.Errorf("get entity model failed, err:%s", entityModelErr.Error())
+		return
+	}
+
+	entityModelVal, entityModelErr = o1.Insert(entityModelVal)
+	if entityModelErr != nil {
+		t.Errorf("insert entity model failed, err:%s", entityModelErr.Error())
+		return
+	}
+
+	newEntityPtr := entityModelVal.Interface(true).(*Entity)
+	if newEntityPtr.Namespace != "test" || newEntityPtr.EType != "test" || newEntityPtr.EName != "test" {
+		t.Errorf("insert entity model failed")
+		return
+	}
+
+	onlineEntityPtr := &OnlineEntity{
+		SessionID:   "abc",
+		Entity:      newEntityPtr,
+		RefreshTime: 10000,
+		ExpireTime:  20000,
+		Namespace:   "test",
+	}
+
+	onlineEntityModelVal, onlineEntityModelErr := localProvider.GetEntityModel(onlineEntityPtr)
+	if onlineEntityModelErr != nil {
+		t.Errorf("get online entity model failed, err:%s", onlineEntityModelErr.Error())
+		return
+	}
+	onlineEntityModelVal, onlineEntityModelErr = o1.Insert(onlineEntityModelVal)
+	if onlineEntityModelErr != nil {
+		t.Errorf("insert online entity failed, err:%s", onlineEntityModelErr.Error())
+		return
+	}
+	newOnlineEntityPtr := onlineEntityModelVal.Interface(true).(*OnlineEntity)
+	if newOnlineEntityPtr.SessionID != onlineEntityPtr.SessionID || newOnlineEntityPtr.RefreshTime != onlineEntityPtr.RefreshTime || newOnlineEntityPtr.ExpireTime != onlineEntityPtr.ExpireTime {
+		t.Errorf("insert online entity failed")
+		return
+	}
+
+	queryOnlineEntityPtr := &OnlineEntity{
+		SessionID: onlineEntityPtr.SessionID,
+		Entity:    &Entity{},
+	}
+
+	queryModelVal, queryModelErr := localProvider.GetEntityModel(queryOnlineEntityPtr)
+	if queryModelErr != nil {
+		t.Errorf("query online entity failed, error:%v", queryModelErr)
+		return
+	}
+
+	resultVal, resultErr := o1.Query(queryModelVal)
+	if resultErr != nil {
+		t.Errorf("query online entity failed, error:%v", resultErr)
+		return
+	}
+
+	qVal := resultVal.Interface(true).(*OnlineEntity)
+	if qVal.ID != newOnlineEntityPtr.ID {
+		t.Errorf("query online entity failed")
 		return
 	}
 }
