@@ -1,50 +1,49 @@
+//go:build remote || all
+// +build remote all
+
 package test
 
 import (
-	"github.com/muidea/magicOrm/orm"
-	"github.com/muidea/magicOrm/provider"
-	"github.com/muidea/magicOrm/provider/remote"
 	"testing"
 	"time"
+
+	"github.com/muidea/magicCommon/foundation/util"
+	"github.com/muidea/magicOrm/orm"
+	"github.com/muidea/magicOrm/provider"
+	"github.com/muidea/magicOrm/provider/helper"
+	"github.com/muidea/magicOrm/provider/remote"
 )
 
 const remoteOwner = "remote"
 
 func TestRemoteSimple(t *testing.T) {
 	orm.Initialize()
-	defer orm.Uninitialize()
+	defer orm.Uninitialized()
 
-	config := orm.NewConfig("localhost:3306", "testdb", "root", "rootkit")
+	config := orm.NewConfig("localhost:3306", "testdb", "root", "rootkit", "")
 	remoteProvider := provider.NewRemoteProvider(remoteOwner)
 
-	o1, err := orm.NewOrm(remoteProvider, config)
+	o1, err := orm.NewOrm(remoteProvider, config, "abc")
 	defer o1.Release()
 	if err != nil {
 		t.Errorf("new Orm failed, err:%s", err.Error())
 		return
 	}
 
-	reference := &Simple{}
-	simpleDef, simpleErr := remote.GetObject(reference)
-	if simpleErr != nil {
-		t.Errorf("GetObject failed, err:%s", simpleErr.Error())
+	objList := []any{&Simple{}}
+	mList, mErr := registerRemoteModel(remoteProvider, objList)
+	if mErr != nil {
+		t.Errorf("registerRemoteModel failed, err:%s", mErr.Error())
 		return
 	}
 
-	objList := []interface{}{simpleDef}
-	_, err = registerModel(remoteProvider, objList)
-	if err != nil {
-		t.Errorf("registerModel failed, err:%s", err.Error())
-		return
-	}
-
-	err = o1.Drop(simpleDef)
+	err = dropModel(o1, mList)
 	if err != nil {
 		t.Errorf("drop reference schema failed, err:%s", err.Error())
 		return
 	}
 
-	err = o1.Create(simpleDef)
+	err = createModel(o1, mList)
 	if err != nil {
 		t.Errorf("create reference schema failed, err:%s", err.Error())
 		return
@@ -59,6 +58,17 @@ func TestRemoteSimple(t *testing.T) {
 		return
 	}
 
+	byteVal, byteErr := remote.EncodeObjectValue(s1Value)
+	if byteErr != nil {
+		t.Errorf("encode object value failed, err:%s", byteErr.Error())
+		return
+	}
+
+	s1Value, s1Err = remote.DecodeObjectValue(byteVal)
+	if s1Err != nil {
+		t.Errorf("decode object value failed, err:%s", s1Err.Error())
+		return
+	}
 	s1Model, s1Err := remoteProvider.GetEntityModel(s1Value)
 	if s1Err != nil {
 		t.Errorf("GetEntityModel failed, err:%s", s1Err.Error())
@@ -70,7 +80,7 @@ func TestRemoteSimple(t *testing.T) {
 		t.Errorf("insert reference failed, err:%s", err.Error())
 		return
 	}
-	err = provider.UpdateEntity(s1Model.Interface(true).(*remote.ObjectValue), s1)
+	err = helper.UpdateEntity(s1Model.Interface(true).(*remote.ObjectValue), s1)
 	if err != nil {
 		t.Errorf("updateEntity failed, err:%s", err.Error())
 		return
@@ -94,7 +104,7 @@ func TestRemoteSimple(t *testing.T) {
 		t.Errorf("update reference failed, err:%s", err.Error())
 		return
 	}
-	err = provider.UpdateEntity(s1Model.Interface(true).(*remote.ObjectValue), s1)
+	err = helper.UpdateEntity(s1Model.Interface(true).(*remote.ObjectValue), s1)
 	if err != nil {
 		t.Errorf("updateEntity failed, err:%s", err.Error())
 		return
@@ -117,7 +127,7 @@ func TestRemoteSimple(t *testing.T) {
 		t.Errorf("query reference failed, err:%s", err.Error())
 		return
 	}
-	err = provider.UpdateEntity(s2Model.Interface(true).(*remote.ObjectValue), s2)
+	err = helper.UpdateEntity(s2Model.Interface(true).(*remote.ObjectValue), s2)
 	if err != nil {
 		t.Errorf("updateEntity failed, err:%s", err.Error())
 		return
@@ -130,45 +140,38 @@ func TestRemoteSimple(t *testing.T) {
 
 func TestRemoteReference(t *testing.T) {
 	orm.Initialize()
-	defer orm.Uninitialize()
+	defer orm.Uninitialized()
 
-	config := orm.NewConfig("localhost:3306", "testdb", "root", "rootkit")
+	config := orm.NewConfig("localhost:3306", "testdb", "root", "rootkit", "")
 	remoteProvider := provider.NewRemoteProvider(remoteOwner)
 
-	o1, err := orm.NewOrm(remoteProvider, config)
+	o1, err := orm.NewOrm(remoteProvider, config, "abc")
 	defer o1.Release()
 	if err != nil {
 		t.Errorf("new Orm failed, err:%s", err.Error())
 		return
 	}
 
-	ref := &Reference{}
-	refDef, refErr := remote.GetObject(ref)
-	if refErr != nil {
-		t.Errorf("GetObject failed, err:%s", refErr.Error())
+	objList := []any{&Reference{}}
+	mList, mErr := registerRemoteModel(remoteProvider, objList)
+	if mErr != nil {
+		t.Errorf("register model failed. err:%s", mErr.Error())
 		return
 	}
 
-	objList := []interface{}{refDef}
-	_, err = registerModel(remoteProvider, objList)
-	if err != nil {
-		t.Errorf("register model failed. err:%s", err.Error())
-		return
-	}
-
-	err = o1.Drop(refDef)
+	err = dropModel(o1, mList)
 	if err != nil {
 		t.Errorf("drop reference schema failed, err:%s", err.Error())
 		return
 	}
 
-	err = o1.Create(refDef)
+	err = createModel(o1, mList)
 	if err != nil {
 		t.Errorf("create reference schema failed, err:%s", err.Error())
 		return
 	}
 
-	ts, _ := time.ParseInLocation("2006-01-02 15:04:05:0000", "2018-01-02 15:04:05:0000", time.Local)
+	ts, _ := time.ParseInLocation(util.CSTLayout, "2018-01-02 15:04:05", time.Local)
 	strValue := "test code"
 	fValue := float32(12.34)
 	flag := true
@@ -176,13 +179,13 @@ func TestRemoteReference(t *testing.T) {
 	fArray := []float32{12.34, 23, 45, 45, 67}
 	strArray := []string{"Abc", "Bcd"}
 	bArray := []bool{true, true, false, false}
-	strPtrArray := []*string{&strValue, &strValue}
+	strPtrArray := []string{strValue, strValue}
 	s1 := &Reference{
 		Name:        strValue,
-		FValue:      &fValue,
+		FValue:      fValue,
 		F64:         23.456,
-		TimeStamp:   &ts,
-		Flag:        &flag,
+		TimeStamp:   ts,
+		Flag:        flag,
 		IArray:      iArray,
 		FArray:      fArray,
 		StrArray:    strArray,
@@ -210,7 +213,7 @@ func TestRemoteReference(t *testing.T) {
 		t.Errorf("insert reference failed, err:%s", err.Error())
 		return
 	}
-	err = provider.UpdateEntity(s1Model.Interface(true).(*remote.ObjectValue), s1)
+	err = helper.UpdateEntity(s1Model.Interface(true).(*remote.ObjectValue), s1)
 	if err != nil {
 		t.Errorf("updateEntity failed, err:%s", err.Error())
 		return
@@ -233,7 +236,7 @@ func TestRemoteReference(t *testing.T) {
 		t.Errorf("update reference failed, err:%s", err.Error())
 		return
 	}
-	err = provider.UpdateEntity(s1Model.Interface(true).(*remote.ObjectValue), s1)
+	err = helper.UpdateEntity(s1Model.Interface(true).(*remote.ObjectValue), s1)
 	if err != nil {
 		t.Errorf("updateEntity failed, err:%s", err.Error())
 		return
@@ -243,18 +246,18 @@ func TestRemoteReference(t *testing.T) {
 	var ts2 time.Time
 	var bVal bool
 	strArray2 := []string{}
-	ptrArray2 := []*string{}
+	ptrArray2 := []string{}
 	s2 := &Reference{
 		ID:          s1.ID,
-		FValue:      &fValue2,
-		TimeStamp:   &ts2,
-		Flag:        &bVal,
+		FValue:      fValue2,
+		TimeStamp:   ts2,
+		Flag:        bVal,
 		IArray:      []int{},
 		FArray:      []float32{},
 		StrArray:    []string{},
 		BArray:      []bool{},
 		PtrArray:    &strArray2,
-		StrPtrArray: []*string{},
+		StrPtrArray: []string{},
 		PtrStrArray: &ptrArray2,
 	}
 	s2Value, s2Err := getObjectValue(s2)
@@ -273,7 +276,7 @@ func TestRemoteReference(t *testing.T) {
 		t.Errorf("query reference failed, err:%s", err.Error())
 		return
 	}
-	err = provider.UpdateEntity(s2Model.Interface(true).(*remote.ObjectValue), s2)
+	err = helper.UpdateEntity(s2Model.Interface(true).(*remote.ObjectValue), s2)
 	if err != nil {
 		t.Errorf("updateEntity failed, err:%s", err.Error())
 		return
@@ -284,7 +287,8 @@ func TestRemoteReference(t *testing.T) {
 	}
 
 	s4 := &Reference{
-		ID: s1.ID,
+		ID:     s1.ID,
+		IArray: []int{},
 	}
 	s4Value, s4Err := getObjectValue(s4)
 	if s4Err != nil {
@@ -302,7 +306,7 @@ func TestRemoteReference(t *testing.T) {
 		t.Errorf("query reference failed, err:%s", err.Error())
 		return
 	}
-	err = provider.UpdateEntity(s4Model.Interface(true).(*remote.ObjectValue), s4)
+	err = helper.UpdateEntity(s4Model.Interface(true).(*remote.ObjectValue), s4)
 	if err != nil {
 		t.Errorf("updateEntity failed, err:%s", err.Error())
 		return
@@ -311,67 +315,47 @@ func TestRemoteReference(t *testing.T) {
 		t.Errorf("query reference failed, err:%s", err.Error())
 		return
 	}
-	if s4.FValue != nil || s4.TimeStamp != nil || s4.Flag != nil || s4.PtrStrArray != nil || s4.PtrArray != nil {
-		t.Errorf("query reference failed, err:%s", err.Error())
+	if s4.IArray == nil || s4.PtrStrArray != nil || s4.PtrArray != nil {
+		t.Errorf("query reference failed")
 		return
 	}
 }
 
 func TestRemoteCompose(t *testing.T) {
 	orm.Initialize()
-	defer orm.Uninitialize()
+	defer orm.Uninitialized()
 
-	config := orm.NewConfig("localhost:3306", "testdb", "root", "rootkit")
+	config := orm.NewConfig("localhost:3306", "testdb", "root", "rootkit", "")
 	remoteProvider := provider.NewRemoteProvider(remoteOwner)
 
-	o1, err := orm.NewOrm(remoteProvider, config)
+	o1, err := orm.NewOrm(remoteProvider, config, "abc")
 	defer o1.Release()
 	if err != nil {
 		t.Errorf("new Orm failed, err:%s", err.Error())
 		return
 	}
 
-	simpleDef, simpleErr := remote.GetObject(&Simple{})
-	if simpleErr != nil {
-		t.Errorf("GetObject failed, err:%s", simpleErr.Error())
-		return
-	}
-
-	referenceDef, referenceErr := remote.GetObject(&Reference{})
-	if referenceErr != nil {
-		t.Errorf("GetObject failed, err:%s", referenceErr.Error())
-		return
-	}
-
-	composeDef, composeErr := remote.GetObject(&Compose{})
-	if composeErr != nil {
-		t.Errorf("GetObject failed, err:%s", composeErr.Error())
-		return
-	}
-
-	objList := []interface{}{simpleDef, referenceDef, composeDef}
-	mList, mErr := registerModel(remoteProvider, objList)
+	objList := []any{&Simple{}, &Reference{}, &Compose{}}
+	mList, mErr := registerRemoteModel(remoteProvider, objList)
 	if mErr != nil {
 		err = mErr
 		t.Errorf("register model failed. err:%s", err.Error())
 		return
 	}
 
-	for _, val := range mList {
-		err = o1.Drop(val)
-		if err != nil {
-			t.Errorf("drop object failed, err:%s", err.Error())
-		}
+	err = dropModel(o1, mList)
+	if err != nil {
+		t.Errorf("drop model failed. err:%s", err.Error())
+		return
 	}
 
-	for _, val := range mList {
-		err = o1.Create(val)
-		if err != nil {
-			t.Errorf("create object failed, err:%s", err.Error())
-		}
+	err = createModel(o1, mList)
+	if err != nil {
+		t.Errorf("create model failed. err:%s", err.Error())
+		return
 	}
 
-	ts, _ := time.Parse("2006-01-02 15:04:05", "2018-01-02 15:04:05")
+	ts, _ := time.Parse(util.CSTLayout, "2018-01-02 15:04:05")
 	s1 := &Simple{I8: 12, I16: 23, I32: 34, I64: 45, Name: "test code", Value: 12.345, F64: 23.456, TimeStamp: ts, Flag: true}
 
 	s1Val, s1Err := getObjectValue(s1)
@@ -391,7 +375,7 @@ func TestRemoteCompose(t *testing.T) {
 		return
 	}
 	s1Val = s1Model.Interface(true).(*remote.ObjectValue)
-	err = provider.UpdateEntity(s1Val, s1)
+	err = helper.UpdateEntity(s1Val, s1)
 	if err != nil {
 		t.Errorf("UpdateEntity failed, err:%s", err.Error())
 		return
@@ -404,13 +388,13 @@ func TestRemoteCompose(t *testing.T) {
 	fArray := []float32{12.34, 23, 45, 45, 67}
 	strArray := []string{"Abc", "Bcd"}
 	bArray := []bool{true, true, false, false}
-	strPtrArray := []*string{&strValue, &strValue}
+	strPtrArray := []string{strValue, strValue}
 	r1 := &Reference{
 		Name:        strValue,
-		FValue:      &fValue,
+		FValue:      fValue,
 		F64:         23.456,
-		TimeStamp:   &ts,
-		Flag:        &flag,
+		TimeStamp:   ts,
+		Flag:        flag,
 		IArray:      iArray,
 		FArray:      fArray,
 		StrArray:    strArray,
@@ -438,7 +422,7 @@ func TestRemoteCompose(t *testing.T) {
 	}
 
 	r1Val = r1Model.Interface(true).(*remote.ObjectValue)
-	err = provider.UpdateEntity(r1Val, r1)
+	err = helper.UpdateEntity(r1Val, r1)
 	if err != nil {
 		t.Errorf("UpdateEntity failed, err:%s", err.Error())
 		return
@@ -446,16 +430,15 @@ func TestRemoteCompose(t *testing.T) {
 
 	refPtrArray := []*Reference{r1}
 	c1 := &Compose{
-		Name:           strValue,
-		Simple:         *s1,
-		PtrSimple:      s1,
-		SimpleArray:    []Simple{*s1, *s1},
-		SimplePtrArray: []*Simple{s1, s1},
-		Reference:      *r1,
-		PtrReference:   r1,
-		RefArray:       []Reference{*r1, *r1, *r1},
-		RefPtrArray:    refPtrArray,
-		PtrRefArray:    &refPtrArray,
+		Name:              strValue,
+		Simple:            *s1,
+		SimplePtr:         s1,
+		SimpleArray:       []Simple{*s1, *s1},
+		SimplePtrArray:    []*Simple{s1, s1},
+		Reference:         *r1,
+		ReferencePtr:      r1,
+		ReferenceArray:    []Reference{*r1, *r1, *r1},
+		ReferencePtrArray: refPtrArray,
 	}
 	c1Val, c1Err := getObjectValue(c1)
 	if c1Err != nil {
@@ -473,25 +456,25 @@ func TestRemoteCompose(t *testing.T) {
 		return
 	}
 	c1Val = c1Model.Interface(true).(*remote.ObjectValue)
-	err = provider.UpdateEntity(c1Val, c1)
+
+	err = helper.UpdateEntity(c1Val, c1)
 	if err != nil {
 		t.Errorf("UpdateEntity failed, err:%s", err.Error())
 		return
 	}
 
 	c2 := &Compose{
-		Name:           strValue,
-		Simple:         *s1,
-		PtrSimple:      s1,
-		SimpleArray:    []Simple{*s1, *s1},
-		SimplePtrArray: []*Simple{s1, s1},
-		PtrSimpleArray: &[]Simple{},
-		Reference:      *r1,
-		PtrReference:   r1,
-		RefArray:       []Reference{*r1, *r1, *r1},
-		RefPtrArray:    refPtrArray,
-		PtrRefArray:    &refPtrArray,
-		PtrCompose:     c1,
+		Name:              strValue,
+		Simple:            *s1,
+		SimplePtr:         s1,
+		SimpleArray:       []Simple{*s1, *s1},
+		SimplePtrArray:    []*Simple{s1, s1},
+		SimpleArrayPtr:    &[]Simple{},
+		Reference:         *r1,
+		ReferencePtr:      r1,
+		ReferenceArray:    []Reference{*r1, *r1, *r1},
+		ReferencePtrArray: refPtrArray,
+		ComposePtr:        c1,
 	}
 	c2Val, c2Err := getObjectValue(c2)
 	if c2Err != nil {
@@ -510,22 +493,22 @@ func TestRemoteCompose(t *testing.T) {
 		return
 	}
 	c2Val = c2Model.Interface(true).(*remote.ObjectValue)
-	err = provider.UpdateEntity(c2Val, c2)
+	err = helper.UpdateEntity(c2Val, c2)
 	if err != nil {
 		t.Errorf("UpdateEntity failed, err:%s", err.Error())
 		return
 	}
 
 	c3 := &Compose{
-		ID:             c2.ID,
-		PtrSimple:      &Simple{},
-		SimpleArray:    []Simple{},
-		SimplePtrArray: []*Simple{},
-		PtrSimpleArray: &[]Simple{},
-		PtrReference:   &Reference{},
-		RefArray:       []Reference{},
-		PtrRefArray:    &[]*Reference{},
-		PtrCompose:     &Compose{},
+		ID:                c2.ID,
+		SimplePtr:         &Simple{},
+		SimpleArray:       []Simple{},
+		SimplePtrArray:    []*Simple{},
+		SimpleArrayPtr:    &[]Simple{},
+		ReferencePtr:      &Reference{},
+		ReferenceArray:    []Reference{},
+		ReferencePtrArray: []*Reference{},
+		ComposePtr:        &Compose{},
 	}
 	c3Val, c3Err := getObjectValue(c3)
 	if c3Err != nil {
@@ -544,7 +527,7 @@ func TestRemoteCompose(t *testing.T) {
 		return
 	}
 	c3Val = c3Model.Interface(true).(*remote.ObjectValue)
-	err = provider.UpdateEntity(c3Val, c3)
+	err = helper.UpdateEntity(c3Val, c3)
 	if err != nil {
 		t.Errorf("UpdateEntity failed, err:%s", err.Error())
 		return
@@ -562,12 +545,12 @@ func TestRemoteCompose(t *testing.T) {
 
 func TestRemoteQuery(t *testing.T) {
 	orm.Initialize()
-	defer orm.Uninitialize()
+	defer orm.Uninitialized()
 
-	config := orm.NewConfig("localhost:3306", "testdb", "root", "rootkit")
+	config := orm.NewConfig("localhost:3306", "testdb", "root", "rootkit", "")
 	remoteProvider := provider.NewRemoteProvider(remoteOwner)
 
-	o1, err := orm.NewOrm(remoteProvider, config)
+	o1, err := orm.NewOrm(remoteProvider, config, "abc")
 	defer o1.Release()
 	if err != nil {
 		t.Errorf("new Orm failed, err:%s", err.Error())
@@ -575,28 +558,28 @@ func TestRemoteQuery(t *testing.T) {
 	}
 
 	s1 := &Simple{}
-	s1Def, s1Err := remote.GetObject(s1)
+	s1Def, s1Err := helper.GetObject(s1)
 	if s1Err != nil {
 		t.Errorf("GetObject failed, err:%s", s1Err.Error())
 		return
 	}
 
 	r1 := &Reference{}
-	r1Def, r1Err := remote.GetObject(r1)
+	r1Def, r1Err := helper.GetObject(r1)
 	if r1Err != nil {
 		t.Errorf("GetObject failed, err:%s", r1Err.Error())
 		return
 	}
 
 	c1 := &Compose{}
-	c1Def, c1Err := remote.GetObject(c1)
+	c1Def, c1Err := helper.GetObject(c1)
 	if c1Err != nil {
 		t.Errorf("GetObject failed, err:%s", c1Err.Error())
 		return
 	}
 
-	objList := []interface{}{s1Def, r1Def, c1Def}
-	mList, mErr := registerModel(remoteProvider, objList)
+	objList := []any{s1Def, r1Def, c1Def}
+	mList, mErr := registerLocalModel(remoteProvider, objList)
 	if mErr != nil {
 		t.Errorf("register model failed. err:%s", mErr.Error())
 		return
@@ -635,7 +618,7 @@ func TestRemoteQuery(t *testing.T) {
 		t.Errorf("insert simple failed, err:%s", err.Error())
 		return
 	}
-	err = provider.UpdateEntity(s1Model.Interface(true).(*remote.ObjectValue), s1)
+	err = helper.UpdateEntity(s1Model.Interface(true).(*remote.ObjectValue), s1)
 	if err != nil {
 		t.Errorf("updateEntity failed, err:%s", err.Error())
 		return
@@ -648,13 +631,13 @@ func TestRemoteQuery(t *testing.T) {
 	fArray := []float32{12.34, 23, 45, 45, 67}
 	strArray := []string{"Abc", "Bcd"}
 	bArray := []bool{true, true, false, false}
-	strPtrArray := []*string{&strValue, &strValue}
+	strPtrArray := []string{strValue, strValue}
 	r1 = &Reference{
 		Name:        strValue,
-		FValue:      &fValue,
+		FValue:      fValue,
 		F64:         23.456,
-		TimeStamp:   &ts,
-		Flag:        &flag,
+		TimeStamp:   ts,
+		Flag:        flag,
 		IArray:      iArray,
 		FArray:      fArray,
 		StrArray:    strArray,
@@ -674,11 +657,11 @@ func TestRemoteQuery(t *testing.T) {
 		return
 	}
 	r1Model, r1Err = o1.Insert(r1Model)
-	if err != nil {
+	if r1Err != nil {
 		t.Errorf("insert reference failed, err:%s", err.Error())
 		return
 	}
-	err = provider.UpdateEntity(r1Model.Interface(true).(*remote.ObjectValue), r1)
+	err = helper.UpdateEntity(r1Model.Interface(true).(*remote.ObjectValue), r1)
 	if err != nil {
 		t.Errorf("updateEntity failed, err:%s", err.Error())
 		return
@@ -686,16 +669,15 @@ func TestRemoteQuery(t *testing.T) {
 
 	refPtrArray := []*Reference{r1}
 	c1 = &Compose{
-		Name:           strValue,
-		Simple:         *s1,
-		PtrSimple:      s1,
-		SimpleArray:    []Simple{*s1, *s1},
-		SimplePtrArray: []*Simple{s1, s1},
-		Reference:      *r1,
-		PtrReference:   r1,
-		RefArray:       []Reference{*r1, *r1, *r1},
-		RefPtrArray:    refPtrArray,
-		PtrRefArray:    &refPtrArray,
+		Name:              strValue,
+		Simple:            *s1,
+		SimplePtr:         s1,
+		SimpleArray:       []Simple{*s1, *s1},
+		SimplePtrArray:    []*Simple{s1, s1},
+		Reference:         *r1,
+		ReferencePtr:      r1,
+		ReferenceArray:    []Reference{*r1, *r1, *r1},
+		ReferencePtrArray: refPtrArray,
 	}
 	c1Value, c1Err := getObjectValue(c1)
 	if c1Err != nil {
@@ -714,7 +696,7 @@ func TestRemoteQuery(t *testing.T) {
 		t.Errorf("insert compose failed, err:%s", err.Error())
 		return
 	}
-	err = provider.UpdateEntity(c1Model.Interface(true).(*remote.ObjectValue), c1)
+	err = helper.UpdateEntity(c1Model.Interface(true).(*remote.ObjectValue), c1)
 	if err != nil {
 		t.Errorf("updateEntity failed, err:%s", err.Error())
 		return
@@ -722,17 +704,16 @@ func TestRemoteQuery(t *testing.T) {
 
 	strValue = "123"
 	c2 := &Compose{
-		Name:           strValue,
-		Simple:         *s1,
-		PtrSimple:      s1,
-		SimpleArray:    []Simple{*s1, *s1},
-		SimplePtrArray: []*Simple{s1, s1},
-		Reference:      *r1,
-		PtrReference:   r1,
-		RefArray:       []Reference{*r1, *r1, *r1},
-		RefPtrArray:    refPtrArray,
-		PtrRefArray:    &refPtrArray,
-		PtrCompose:     c1,
+		Name:              strValue,
+		Simple:            *s1,
+		SimplePtr:         s1,
+		SimpleArray:       []Simple{*s1, *s1},
+		SimplePtrArray:    []*Simple{s1, s1},
+		Reference:         *r1,
+		ReferencePtr:      r1,
+		ReferenceArray:    []Reference{*r1, *r1, *r1},
+		ReferencePtrArray: refPtrArray,
+		ComposePtr:        c1,
 	}
 	c2Value, c2Err := getObjectValue(c2)
 	if c2Err != nil {
@@ -750,13 +731,13 @@ func TestRemoteQuery(t *testing.T) {
 		t.Errorf("insert compose failed, err:%s", err.Error())
 		return
 	}
-	err = provider.UpdateEntity(c2Model.Interface(true).(*remote.ObjectValue), c2)
+	err = helper.UpdateEntity(c2Model.Interface(true).(*remote.ObjectValue), c2)
 	if err != nil {
 		t.Errorf("updateEntity failed, err:%s", err.Error())
 		return
 	}
 
-	c3 := &Compose{ID: c2.ID, PtrSimple: &Simple{}, PtrSimpleArray: &[]Simple{}, PtrReference: &Reference{}, PtrRefArray: &[]*Reference{}, PtrCompose: &Compose{}}
+	c3 := &Compose{ID: c2.ID, SimplePtr: &Simple{}, SimpleArrayPtr: &[]Simple{}, ReferencePtr: &Reference{}, ComposePtr: &Compose{}}
 	c3Value := c2Value
 	c3Model, c3Err := remoteProvider.GetEntityModel(c3Value)
 	if c3Err != nil {
@@ -770,7 +751,7 @@ func TestRemoteQuery(t *testing.T) {
 		return
 	}
 
-	err = provider.UpdateEntity(c3Model.Interface(true).(*remote.ObjectValue), c3)
+	err = helper.UpdateEntity(c3Model.Interface(true).(*remote.ObjectValue), c3)
 	if err != nil {
 		t.Errorf("updateEntity failed, err:%s", err.Error())
 		return
@@ -790,25 +771,32 @@ func TestRemoteQuery(t *testing.T) {
 		return
 	}
 
-	err = provider.UpdateEntity(c4Model.Interface(true).(*remote.ObjectValue), c4)
+	err = helper.UpdateEntity(c4Model.Interface(true).(*remote.ObjectValue), c4)
 	if err != nil {
 		t.Errorf("updateEntity failed, err:%s", err.Error())
 		return
 	}
 
-	cList := []*Compose{}
-	cListValue, cListErr := remote.GetSliceObjectValue(&cList)
-	if cListErr != nil {
-		t.Errorf("getSliceObjectValue failed, err:%s", cListErr.Error())
+	cComposePtr := &Compose{}
+	cObjectValue, cObjectValueErr := helper.GetObjectValue(cComposePtr)
+	if cObjectValueErr != nil {
+		t.Errorf("GetObjectValue failed, err:%s", cObjectValueErr.Error())
 		return
 	}
-	cListModel, cListErr := remoteProvider.GetEntityModel(cListValue)
-	if cListErr != nil {
-		t.Errorf("GetEntityModel failed, err:%s", cListErr.Error())
+	cObjectPtr, cObjectErr := helper.GetObject(cComposePtr)
+	if cObjectErr != nil {
+		t.Errorf("GetObject failed, err:%s", cObjectErr.Error())
 		return
 	}
 
-	cModelList, cModelErr := o1.BatchQuery(cListModel, nil)
+	filter, err := remoteProvider.GetModelFilter(cObjectPtr)
+	if err != nil {
+		t.Errorf("GetEntityFilter failed, err:%s", err.Error())
+		return
+	}
+
+	filter.ValueMask(cObjectValue)
+	cModelList, cModelErr := o1.BatchQuery(filter)
 	if cModelErr != nil {
 		err = cModelErr
 		t.Errorf("batch query compose failed, err:%s", err.Error())
@@ -819,23 +807,15 @@ func TestRemoteQuery(t *testing.T) {
 		return
 	}
 
-	cList = []*Compose{}
-	cListValue, cListErr = remote.GetSliceObjectValue(&cList)
-	if cListErr != nil {
-		t.Errorf("getSliceObjectValue failed, err:%s", c2Err.Error())
-		return
-	}
-
-	maskVal, maskErr := remote.GetObjectValue(&Compose{PtrSimple: &Simple{}})
+	maskVal, maskErr := helper.GetObjectValue(&Compose{SimplePtr: &Simple{}})
 	if maskErr != nil {
 		t.Errorf("getObjectValue failed, err:%s", maskErr.Error())
 		return
 	}
 
-	filter := orm.GetFilter(remoteProvider)
 	filter.Equal("name", strValue)
 	filter.ValueMask(maskVal)
-	cModelList, cModelErr = o1.BatchQuery(cListModel, filter)
+	cModelList, cModelErr = o1.BatchQuery(filter)
 	if cModelErr != nil {
 		t.Errorf("batch query compose failed, err:%s", cModelErr.Error())
 		return
@@ -843,5 +823,122 @@ func TestRemoteQuery(t *testing.T) {
 	if len(cModelList) != 3 {
 		t.Errorf("batch query compose failed")
 		return
+	}
+}
+
+func TestRemoteOnlineEntity(t *testing.T) {
+	orm.Initialize()
+	defer orm.Uninitialized()
+
+	config := orm.NewConfig("localhost:3306", "testdb", "root", "rootkit", "")
+	remoteProvider := provider.NewRemoteProvider(remoteOwner)
+
+	o1, err := orm.NewOrm(remoteProvider, config, "abc")
+	defer o1.Release()
+	if err != nil {
+		t.Errorf("new Orm failed, err:%s", err.Error())
+		return
+	}
+
+	objList := []any{&Entity{}, &OnlineEntity{}}
+	mList, mErr := registerRemoteModel(remoteProvider, objList)
+	if mErr != nil {
+		t.Errorf("register model failed. err:%s", mErr.Error())
+		return
+	}
+
+	err = dropModel(o1, mList)
+	if err != nil {
+		t.Errorf("drop model failed. err:%s", err.Error())
+		return
+	}
+
+	err = createModel(o1, mList)
+	if err != nil {
+		t.Errorf("create model failed. err:%s", err.Error())
+		return
+	}
+
+	entityPtr := &Entity{EName: "test", EType: "test", EID: 10, Namespace: "test"}
+	remoteEntityVal, remoteEntityErr := getObjectValue(entityPtr)
+	if remoteEntityErr != nil {
+		t.Errorf("getObjectValue failed. err:%s", remoteEntityErr.Error())
+		return
+	}
+
+	entityModelVal, entityModelErr := remoteProvider.GetEntityModel(remoteEntityVal)
+	if entityModelErr != nil {
+		t.Errorf("get entity model failed, err:%s", entityModelErr.Error())
+		return
+	}
+
+	entityModelVal, entityModelErr = o1.Insert(entityModelVal)
+	if entityModelErr != nil {
+		t.Errorf("insert entity model failed, err:%s", entityModelErr.Error())
+		return
+	}
+
+	newRemoteEntityPtr := entityModelVal.Interface(true).(*remote.ObjectValue)
+	if len(newRemoteEntityPtr.Fields) != 5 {
+		t.Errorf("insert entity model failed")
+		return
+	}
+
+	entityPtr.ID = 1
+	onlineEntityPtr := &OnlineEntity{
+		SessionID:   "abc",
+		Entity:      entityPtr,
+		RefreshTime: 10000,
+		ExpireTime:  20000,
+		Namespace:   "test",
+	}
+	remoteOnlineEntityVal, remoteOnlineEntityErr := getObjectValue(onlineEntityPtr)
+	if remoteOnlineEntityErr != nil {
+		t.Errorf("get remote online entity value failed")
+		return
+	}
+
+	onlineEntityModelVal, onlineEntityModelErr := remoteProvider.GetEntityModel(remoteOnlineEntityVal)
+	if onlineEntityModelErr != nil {
+		t.Errorf("get online entity model failed, err:%s", onlineEntityModelErr.Error())
+		return
+	}
+	onlineEntityModelVal, onlineEntityModelErr = o1.Insert(onlineEntityModelVal)
+	if onlineEntityModelErr != nil {
+		t.Errorf("insert online entity failed, err:%s", onlineEntityModelErr.Error())
+		return
+	}
+	newOnlineEntityPtr := onlineEntityModelVal.Interface(true).(*remote.ObjectValue)
+	if len(newOnlineEntityPtr.Fields) != 6 {
+		t.Errorf("insert online entity failed")
+		return
+	}
+
+	queryOnlineEntityPtr := &OnlineEntity{
+		SessionID: onlineEntityPtr.SessionID,
+		Entity:    &Entity{},
+	}
+
+	remoteQueryVal, remoteQueryErr := getObjectValue(queryOnlineEntityPtr)
+	if remoteQueryErr != nil {
+		t.Errorf("getObjectValue failed, error:%s", remoteQueryErr.Error())
+		return
+	}
+
+	queryModelVal, queryModelErr := remoteProvider.GetEntityModel(remoteQueryVal)
+	if queryModelErr != nil {
+		t.Errorf("query online entity failed, error:%v", queryModelErr)
+		return
+	}
+
+	resultVal, resultErr := o1.Query(queryModelVal)
+	if resultErr != nil {
+		t.Errorf("query online entity failed, error:%v", resultErr)
+		return
+	}
+
+	qVal := resultVal.Interface(true).(*remote.ObjectValue)
+	if len(qVal.Fields) != 6 {
+		t.Errorf("query online entity failed")
 	}
 }
