@@ -1,21 +1,26 @@
+//go:build mixed || all
+// +build mixed all
+
 package test
 
 import (
-	"fmt"
 	"testing"
 	"time"
 
-	"github.com/muidea/magicOrm/model"
+	cd "github.com/muidea/magicCommon/def"
+	"github.com/muidea/magicCommon/foundation/util"
+
 	"github.com/muidea/magicOrm/orm"
 	"github.com/muidea/magicOrm/provider"
+	"github.com/muidea/magicOrm/provider/helper"
 	"github.com/muidea/magicOrm/provider/remote"
 )
 
 const composeLocalOwner = "composeLocal"
 const composeRemoteOwner = "composeRemote"
 
-func prepareLocalData(localProvider provider.Provider, orm orm.Orm) (sPtr *Simple, rPtr *Reference, cPtr *Compose, err error) {
-	ts, _ := time.Parse("2006-01-02 15:04:05", "2018-01-02 15:04:05")
+func prepareLocalData(localProvider provider.Provider, orm orm.Orm) (sPtr *Simple, rPtr *Reference, cPtr *Compose, err *cd.Error) {
+	ts, _ := time.Parse(util.CSTLayout, "2018-01-02 15:04:05")
 	sVal := &Simple{I8: 12, I16: 23, I32: 34, I64: 45, Name: "test code", Value: 12.345, F64: 23.456, TimeStamp: ts, Flag: true}
 
 	sModel, sErr := localProvider.GetEntityModel(sVal)
@@ -38,13 +43,13 @@ func prepareLocalData(localProvider provider.Provider, orm orm.Orm) (sPtr *Simpl
 	fArray := []float32{12.34, 23, 45, 45, 67}
 	strArray := []string{"Abc", "Bcd"}
 	bArray := []bool{true, true, false, false}
-	strPtrArray := []*string{&strValue, &strValue}
+	strPtrArray := []string{strValue, strValue}
 	rVal := &Reference{
 		Name:        strValue,
-		FValue:      &fValue,
+		FValue:      fValue,
 		F64:         23.456,
-		TimeStamp:   &ts,
-		Flag:        &flag,
+		TimeStamp:   ts,
+		Flag:        flag,
 		IArray:      iArray,
 		FArray:      fArray,
 		StrArray:    strArray,
@@ -69,16 +74,15 @@ func prepareLocalData(localProvider provider.Provider, orm orm.Orm) (sPtr *Simpl
 
 	refPtrArray := []*Reference{rPtr}
 	cVal := &Compose{
-		Name:           strValue,
-		Simple:         *sPtr,
-		PtrSimple:      sPtr,
-		SimpleArray:    []Simple{*sPtr, *sPtr},
-		SimplePtrArray: []*Simple{sPtr, sPtr},
-		Reference:      *rPtr,
-		PtrReference:   rPtr,
-		RefArray:       []Reference{*rPtr, *rPtr, *rPtr},
-		RefPtrArray:    refPtrArray,
-		PtrRefArray:    &refPtrArray,
+		Name:              strValue,
+		Simple:            *sPtr,
+		SimplePtr:         sPtr,
+		SimpleArray:       []Simple{*sPtr, *sPtr},
+		SimplePtrArray:    []*Simple{sPtr, sPtr},
+		Reference:         *rPtr,
+		ReferencePtr:      rPtr,
+		ReferenceArray:    []Reference{*rPtr, *rPtr, *rPtr},
+		ReferencePtrArray: refPtrArray,
 	}
 	cModel, cErr := localProvider.GetEntityModel(cVal)
 	if cErr != nil {
@@ -96,11 +100,11 @@ func prepareLocalData(localProvider provider.Provider, orm orm.Orm) (sPtr *Simpl
 	return
 }
 
-func prepareRemoteData(remoteProvider provider.Provider, orm orm.Orm) (sPtr *Simple, rPtr *Reference, cPtr *Compose, err error) {
-	ts, _ := time.Parse("2006-01-02 15:04:05", "2018-01-02 15:04:05")
+func prepareRemoteData(remoteProvider provider.Provider, orm orm.Orm) (sPtr *Simple, rPtr *Reference, cPtr *Compose, err *cd.Error) {
+	ts, _ := time.Parse(util.CSTLayout, "2018-01-02 15:04:05")
 	sVal := &Simple{I8: 12, I16: 23, I32: 34, I64: 45, Name: "test code", Value: 12.345, F64: 23.456, TimeStamp: ts, Flag: true}
 
-	sObjectVal, _ := remote.GetObjectValue(sVal)
+	sObjectVal, _ := helper.GetObjectValue(sVal)
 	sModel, sErr := remoteProvider.GetEntityModel(sObjectVal)
 	if sErr != nil {
 		err = sErr
@@ -114,7 +118,7 @@ func prepareRemoteData(remoteProvider provider.Provider, orm orm.Orm) (sPtr *Sim
 	}
 	sObjectVal = sModel.Interface(true).(*remote.ObjectValue)
 	sPtr = &Simple{}
-	sErr = provider.UpdateEntity(sObjectVal, sPtr)
+	sErr = helper.UpdateEntity(sObjectVal, sPtr)
 	if sErr != nil {
 		err = sErr
 		return
@@ -127,13 +131,13 @@ func prepareRemoteData(remoteProvider provider.Provider, orm orm.Orm) (sPtr *Sim
 	fArray := []float32{12.34, 23, 45, 45, 67}
 	strArray := []string{"Abc", "Bcd"}
 	bArray := []bool{true, true, false, false}
-	strPtrArray := []*string{&strValue, &strValue}
+	strPtrArray := []string{strValue, strValue}
 	rVal := &Reference{
 		Name:        strValue,
-		FValue:      &fValue,
+		FValue:      fValue,
 		F64:         23.456,
-		TimeStamp:   &ts,
-		Flag:        &flag,
+		TimeStamp:   ts,
+		Flag:        flag,
 		IArray:      iArray,
 		FArray:      fArray,
 		StrArray:    strArray,
@@ -143,7 +147,7 @@ func prepareRemoteData(remoteProvider provider.Provider, orm orm.Orm) (sPtr *Sim
 		PtrStrArray: &strPtrArray,
 	}
 
-	rObjectVal, _ := remote.GetObjectValue(rVal)
+	rObjectVal, _ := helper.GetObjectValue(rVal)
 	rModel, rErr := remoteProvider.GetEntityModel(rObjectVal)
 	if rErr != nil {
 		err = rErr
@@ -160,10 +164,10 @@ func prepareRemoteData(remoteProvider provider.Provider, orm orm.Orm) (sPtr *Sim
 	var ts2 time.Time
 	var flag2 bool
 	strArray2 := []string{}
-	ptrStrArray := []*string{}
+	ptrStrArray := []string{}
 
-	rPtr = &Reference{FValue: &fVal, TimeStamp: &ts2, Flag: &flag2, PtrArray: &strArray2, PtrStrArray: &ptrStrArray}
-	rErr = provider.UpdateEntity(rObjectVal, rPtr)
+	rPtr = &Reference{FValue: fVal, TimeStamp: ts2, Flag: flag2, PtrArray: &strArray2, PtrStrArray: &ptrStrArray}
+	rErr = helper.UpdateEntity(rObjectVal, rPtr)
 	if rErr != nil {
 		err = rErr
 		return
@@ -171,18 +175,17 @@ func prepareRemoteData(remoteProvider provider.Provider, orm orm.Orm) (sPtr *Sim
 
 	refPtrArray := []*Reference{rPtr}
 	cVal := &Compose{
-		Name:           strValue,
-		Simple:         *sPtr,
-		PtrSimple:      sPtr,
-		SimpleArray:    []Simple{*sPtr, *sPtr},
-		SimplePtrArray: []*Simple{sPtr, sPtr},
-		Reference:      *rPtr,
-		PtrReference:   rPtr,
-		RefArray:       []Reference{*rPtr, *rPtr, *rPtr},
-		RefPtrArray:    refPtrArray,
-		PtrRefArray:    &refPtrArray,
+		Name:              strValue,
+		Simple:            *sPtr,
+		SimplePtr:         sPtr,
+		SimpleArray:       []Simple{*sPtr, *sPtr},
+		SimplePtrArray:    []*Simple{sPtr, sPtr},
+		Reference:         *rPtr,
+		ReferencePtr:      rPtr,
+		ReferenceArray:    []Reference{*rPtr, *rPtr, *rPtr},
+		ReferencePtrArray: refPtrArray,
 	}
-	cObjectVal, _ := remote.GetObjectValue(cVal)
+	cObjectVal, _ := helper.GetObjectValue(cVal)
 	cModel, cErr := remoteProvider.GetEntityModel(cObjectVal)
 	if cErr != nil {
 		err = cErr
@@ -196,17 +199,16 @@ func prepareRemoteData(remoteProvider provider.Provider, orm orm.Orm) (sPtr *Sim
 	}
 	cObjectVal = cModel.Interface(true).(*remote.ObjectValue)
 	cPtr = &Compose{
-		PtrSimple:      &Simple{},
-		SimpleArray:    []Simple{},
-		SimplePtrArray: []*Simple{},
-		PtrSimpleArray: &[]Simple{},
-		PtrReference:   &Reference{},
-		RefArray:       []Reference{},
-		RefPtrArray:    []*Reference{},
-		PtrRefArray:    &[]*Reference{},
-		PtrCompose:     &Compose{},
+		SimplePtr:         &Simple{},
+		SimpleArray:       []Simple{},
+		SimplePtrArray:    []*Simple{},
+		SimpleArrayPtr:    &[]Simple{},
+		ReferencePtr:      &Reference{},
+		ReferenceArray:    []Reference{},
+		ReferencePtrArray: []*Reference{},
+		ComposePtr:        &Compose{},
 	}
-	cErr = provider.UpdateEntity(cObjectVal, cPtr)
+	cErr = helper.UpdateEntity(cObjectVal, cPtr)
 	if cErr != nil {
 		err = cErr
 		return
@@ -217,12 +219,12 @@ func prepareRemoteData(remoteProvider provider.Provider, orm orm.Orm) (sPtr *Sim
 
 func TestComposeLocal(t *testing.T) {
 	orm.Initialize()
-	defer orm.Uninitialize()
+	defer orm.Uninitialized()
 
-	config := orm.NewConfig("localhost:3306", "testdb", "root", "rootkit")
+	config := orm.NewConfig("localhost:3306", "testdb", "root", "rootkit", "")
 	localProvider := provider.NewLocalProvider(composeLocalOwner)
 
-	o1, err := orm.NewOrm(localProvider, config)
+	o1, err := orm.NewOrm(localProvider, config, "abc")
 	defer o1.Release()
 	if err != nil {
 		t.Errorf("new Orm failed, err:%s", err.Error())
@@ -233,8 +235,8 @@ func TestComposeLocal(t *testing.T) {
 	referenceDef := &Reference{}
 	composeDef := &Compose{}
 
-	entityList := []interface{}{simpleDef, referenceDef, composeDef}
-	modelList, modelErr := registerModel(localProvider, entityList)
+	entityList := []any{simpleDef, referenceDef, composeDef}
+	modelList, modelErr := registerLocalModel(localProvider, entityList)
 	if modelErr != nil {
 		err = modelErr
 		t.Errorf("register model failed. err:%s", err.Error())
@@ -253,9 +255,6 @@ func TestComposeLocal(t *testing.T) {
 		return
 	}
 
-	sValList := []*Compose{}
-	sModelList := []model.Model{}
-
 	sPtr, rPtr, cPtr, pErr := prepareLocalData(localProvider, o1)
 	if pErr != nil {
 		t.Errorf("prepareLocalData failed. err:%s", pErr.Error())
@@ -264,180 +263,147 @@ func TestComposeLocal(t *testing.T) {
 
 	strValue := "test code"
 	// insert
-	for idx := 0; idx < 100; idx++ {
-		refPtrArray := []*Reference{rPtr}
-		sVal := &Compose{
-			Name:           strValue,
-			Simple:         *sPtr,
-			PtrSimple:      sPtr,
-			SimpleArray:    []Simple{*sPtr, *sPtr},
-			SimplePtrArray: []*Simple{sPtr, sPtr},
-			Reference:      *rPtr,
-			PtrReference:   rPtr,
-			RefArray:       []Reference{*rPtr, *rPtr, *rPtr},
-			RefPtrArray:    refPtrArray,
-			PtrRefArray:    &refPtrArray,
-			PtrCompose:     cPtr,
-		}
-		sValList = append(sValList, sVal)
-
-		sModel, sErr := localProvider.GetEntityModel(sVal)
-		if sErr != nil {
-			err = sErr
-			t.Errorf("GetEntityModel failed. err:%s", err.Error())
-			return
-		}
-
-		sModelList = append(sModelList, sModel)
+	refPtrArray := []*Reference{rPtr}
+	composePtr := &Compose{
+		Name:              strValue,
+		Simple:            *sPtr,
+		SimplePtr:         sPtr,
+		SimpleArray:       []Simple{*sPtr, *sPtr},
+		SimplePtrArray:    []*Simple{sPtr, sPtr},
+		Reference:         *rPtr,
+		ReferencePtr:      rPtr,
+		ReferenceArray:    []Reference{*rPtr, *rPtr, *rPtr},
+		ReferencePtrArray: refPtrArray,
+		ComposePtr:        cPtr,
 	}
 
-	for idx := 0; idx < 100; idx++ {
-		vModel, vErr := o1.Insert(sModelList[idx])
-		if vErr != nil {
-			err = vErr
-			t.Errorf("Insert failed. err:%s", err.Error())
-			return
-		}
-
-		sModelList[idx] = vModel
-		sValList[idx] = vModel.Interface(true).(*Compose)
-	}
-
-	// update
-	for idx := 0; idx < 100; idx++ {
-		sVal := sValList[idx]
-		sVal.Name = "hi"
-		sModel, sErr := localProvider.GetEntityModel(sVal)
-		if sErr != nil {
-			err = sErr
-			t.Errorf("GetEntityModel failed. err:%s", err.Error())
-			return
-		}
-
-		sModelList[idx] = sModel
-	}
-	for idx := 0; idx < 100; idx++ {
-		vModel, vErr := o1.Update(sModelList[idx])
-		if vErr != nil {
-			err = vErr
-			t.Errorf("Update failed. err:%s", err.Error())
-			return
-		}
-
-		sModelList[idx] = vModel
-		sValList[idx] = vModel.Interface(true).(*Compose)
-	}
-
-	// query
-	qValList := []*Compose{}
-	qModelList := []model.Model{}
-	for idx := 0; idx < 100; idx++ {
-		qVal := &Compose{
-			ID:             sValList[idx].ID,
-			PtrSimple:      &Simple{},
-			SimpleArray:    []Simple{},
-			SimplePtrArray: []*Simple{},
-			PtrSimpleArray: &[]Simple{},
-			PtrReference:   &Reference{},
-			RefArray:       []Reference{},
-			RefPtrArray:    []*Reference{},
-			PtrRefArray:    &[]*Reference{},
-			PtrCompose:     &Compose{},
-		}
-		qValList = append(qValList, qVal)
-
-		qModel, qErr := localProvider.GetEntityModel(qVal)
-		if qErr != nil {
-			err = qErr
-			t.Errorf("GetEntityModel failed. err:%s", err.Error())
-			return
-		}
-
-		qModelList = append(qModelList, qModel)
-	}
-
-	for idx := 0; idx < 100; idx++ {
-		qModel, qErr := o1.Query(qModelList[idx])
-		if qErr != nil {
-			err = qErr
-			t.Errorf("Query failed. err:%s", err.Error())
-			return
-		}
-
-		qModelList[idx] = qModel
-		qValList[idx] = qModel.Interface(true).(*Compose)
-	}
-
-	for idx := 0; idx < 100; idx++ {
-		sVal := sValList[idx]
-		qVal := qValList[idx]
-		if !sVal.IsSame(qVal) {
-			err = fmt.Errorf("compare value failed")
-			t.Errorf("IsSame failed. err:%s", err.Error())
-			return
-		}
-	}
-
-	bqValList := []*Compose{}
-	bqModel, bqErr := localProvider.GetEntityModel(&bqValList)
-	if bqErr != nil {
-		t.Errorf("GetEntityModel failed, err:%s", bqErr.Error())
+	composeModel, composeErr := localProvider.GetEntityModel(composePtr)
+	if composeErr != nil {
+		err = composeErr
+		t.Errorf("GetEntityModel failed. err:%s", err.Error())
 		return
 	}
 
-	filter := orm.GetFilter(localProvider)
-	filter.Equal("name", "hi")
-	filter.ValueMask(&Compose{
-		PtrSimple:      &Simple{},
-		SimpleArray:    []Simple{},
-		SimplePtrArray: []*Simple{},
-		PtrSimpleArray: &[]Simple{},
-		PtrReference:   &Reference{},
-		RefArray:       []Reference{},
-		RefPtrArray:    []*Reference{},
-		PtrRefArray:    &[]*Reference{},
-		PtrCompose:     &Compose{},
+	composeModel, composeErr = o1.Insert(composeModel)
+	if composeErr != nil {
+		err = composeErr
+		t.Errorf("Insert failed. err:%s", err.Error())
+		return
+	}
+
+	// update
+	composePtr = composeModel.Interface(true).(*Compose)
+	composePtr.Name = "hi"
+	composeModel, composeErr = localProvider.GetEntityModel(composePtr)
+	if composeErr != nil {
+		err = composeErr
+		t.Errorf("GetEntityModel failed. err:%s", err.Error())
+		return
+	}
+
+	composeModel, composeErr = o1.Update(composeModel)
+	if composeErr != nil {
+		err = composeErr
+		t.Errorf("Update failed. err:%s", err.Error())
+		return
+	}
+
+	composePtr = composeModel.Interface(true).(*Compose)
+
+	// query
+	queryVal := &Compose{
+		ID:                composePtr.ID,
+		SimplePtr:         &Simple{},
+		SimpleArray:       []Simple{},
+		SimplePtrArray:    []*Simple{},
+		SimpleArrayPtr:    &[]Simple{},
+		ReferencePtr:      &Reference{},
+		ReferenceArray:    []Reference{},
+		ReferencePtrArray: []*Reference{},
+		ComposePtr:        &Compose{},
+	}
+
+	queryModel, queryErr := localProvider.GetEntityModel(queryVal)
+	if queryErr != nil {
+		err = queryErr
+		t.Errorf("GetEntityModel failed. err:%s", err.Error())
+		return
+	}
+
+	queryModel, queryErr = o1.Query(queryModel)
+	if queryErr != nil {
+		err = queryErr
+		t.Errorf("Query failed, err:%s", err.Error())
+		return
+	}
+	queryVal = queryModel.Interface(true).(*Compose)
+
+	if !composePtr.IsSame(queryVal) {
+		err = cd.NewError(cd.Unexpected, "compare value failed")
+		t.Errorf("IsSame failed. err:%s", err.Error())
+		return
+	}
+
+	cModel, _ := localProvider.GetEntityModel(&Compose{})
+	filter, err := localProvider.GetModelFilter(cModel)
+	if err != nil {
+		t.Errorf("GetEntityFilter failed, err:%s", err.Error())
+		return
+	}
+
+	err = filter.Equal("name", "hi")
+	if err != nil {
+		t.Errorf("filter.Equal failed, err:%s", err.Error())
+		return
+	}
+	err = filter.ValueMask(&Compose{
+		SimplePtr:         &Simple{},
+		SimpleArray:       []Simple{},
+		SimplePtrArray:    []*Simple{},
+		SimpleArrayPtr:    &[]Simple{},
+		ReferencePtr:      &Reference{},
+		ReferenceArray:    []Reference{},
+		ReferencePtrArray: []*Reference{},
+		ComposePtr:        &Compose{},
 	})
-	bqModelList, bqModelErr := o1.BatchQuery(bqModel, filter)
+	if err != nil {
+		t.Errorf("filter.ValueMask failed, err:%s", err.Error())
+		return
+	}
+
+	bqModelList, bqModelErr := o1.BatchQuery(filter)
 	if bqModelErr != nil {
 		t.Errorf("BatchQuery failed, err:%s", bqModelErr.Error())
 		return
 	}
-	if len(bqModelList) != 100 {
+	if len(bqModelList) != 1 {
 		t.Errorf("batch query compose failed")
 		return
 	}
 
 	// delete
-	for idx := 0; idx < 100; idx++ {
-		_, qErr := o1.Delete(bqModelList[idx])
-		if qErr != nil {
-			err = qErr
-			return
-		}
+	_, qErr := o1.Delete(bqModelList[0])
+	if qErr != nil {
+		t.Errorf("Delete failed, err:%s", qErr.Error())
+		return
 	}
 }
 
 func TestComposeRemote(t *testing.T) {
-
 	orm.Initialize()
-	defer orm.Uninitialize()
-	config := orm.NewConfig("localhost:3306", "testdb", "root", "rootkit")
+	defer orm.Uninitialized()
+	config := orm.NewConfig("localhost:3306", "testdb", "root", "rootkit", "")
 	remoteProvider := provider.NewRemoteProvider(composeRemoteOwner)
 
-	o1, err := orm.NewOrm(remoteProvider, config)
+	o1, err := orm.NewOrm(remoteProvider, config, "abc")
 	defer o1.Release()
 	if err != nil {
 		t.Errorf("new Orm failed, err:%s", err.Error())
 		return
 	}
 
-	simpleDef, _ := remote.GetObject(&Simple{})
-	referenceDef, _ := remote.GetObject(&Reference{})
-	composeDef, _ := remote.GetObject(&Compose{})
-
-	entityList := []interface{}{simpleDef, referenceDef, composeDef}
-	modelList, modelErr := registerModel(remoteProvider, entityList)
+	entityList := []any{&Simple{}, &Reference{}, &Compose{}}
+	modelList, modelErr := registerRemoteModel(remoteProvider, entityList)
 	if modelErr != nil {
 		err = modelErr
 		t.Errorf("register model failed. err:%s", err.Error())
@@ -462,216 +428,182 @@ func TestComposeRemote(t *testing.T) {
 		return
 	}
 
-	sValList := []*Compose{}
-	sObjectValList := []*remote.ObjectValue{}
-	sModelList := []model.Model{}
-
 	strValue := "test code"
 	// insert
-	for idx := 0; idx < 100; idx++ {
-		refPtrArray := []*Reference{rPtr}
-		sVal := &Compose{
-			Name:           strValue,
-			Simple:         *sPtr,
-			PtrSimple:      sPtr,
-			SimpleArray:    []Simple{*sPtr, *sPtr},
-			SimplePtrArray: []*Simple{sPtr, sPtr},
-			Reference:      *rPtr,
-			PtrReference:   rPtr,
-			RefArray:       []Reference{*rPtr, *rPtr, *rPtr},
-			RefPtrArray:    refPtrArray,
-			PtrRefArray:    &refPtrArray,
-			PtrCompose:     cPtr,
-		}
-		sValList = append(sValList, sVal)
-
-		sObjectVal, sObjectErr := remote.GetObjectValue(sVal)
-		if sObjectErr != nil {
-			err = sObjectErr
-			t.Errorf("GetObjectValue failed. err:%s", err.Error())
-			return
-		}
-		sObjectValList = append(sObjectValList, sObjectVal)
-
-		sModel, sErr := remoteProvider.GetEntityModel(sObjectVal)
-		if sErr != nil {
-			err = sErr
-			t.Errorf("GetEntityModel failed. err:%s", err.Error())
-			return
-		}
-
-		sModelList = append(sModelList, sModel)
+	refPtrArray := []*Reference{rPtr}
+	composePtr := &Compose{
+		Name:              strValue,
+		Simple:            *sPtr,
+		SimplePtr:         sPtr,
+		SimpleArray:       []Simple{*sPtr, *sPtr},
+		SimplePtrArray:    []*Simple{sPtr, sPtr},
+		Reference:         *rPtr,
+		ReferencePtr:      rPtr,
+		ReferenceArray:    []Reference{*rPtr, *rPtr, *rPtr},
+		ReferencePtrArray: refPtrArray,
+		ComposePtr:        cPtr,
+	}
+	composeObjectValue, composeObjectErr := helper.GetObjectValue(composePtr)
+	if composeObjectErr != nil {
+		err = composeObjectErr
+		t.Errorf("GetObjectValue failed. err:%s", err.Error())
+		return
 	}
 
-	for idx := 0; idx < 100; idx++ {
-		vModel, vErr := o1.Insert(sModelList[idx])
-		if vErr != nil {
-			err = vErr
-			t.Errorf("Insert failed. err:%s", err.Error())
-			return
-		}
-
-		sObjectVal := vModel.Interface(true).(*remote.ObjectValue)
-		sVal := sValList[idx]
-		err = provider.UpdateEntity(sObjectVal, sVal)
-		if err != nil {
-			t.Errorf("UpdateEntity failed. err:%s", err.Error())
-			return
-		}
-		sValList[idx] = sVal
-		sModelList[idx] = vModel
-		sObjectValList[idx] = sObjectVal
+	composeModel, composeErr := remoteProvider.GetEntityModel(composeObjectValue)
+	if composeErr != nil {
+		err = composeErr
+		t.Errorf("GetEntityModel failed. err:%s", err.Error())
+		return
 	}
 
-	// update
-	for idx := 0; idx < 100; idx++ {
-		sVal := sValList[idx]
-		sVal.Name = "hi"
-		sObjectVal, sObjectErr := remote.GetObjectValue(sVal)
-		if sObjectErr != nil {
-			err = sObjectErr
-			t.Errorf("GetObjectValue failed. err:%s", err.Error())
-			return
-		}
-		sObjectValList[idx] = sObjectVal
-
-		sModel, sErr := remoteProvider.GetEntityModel(sObjectVal)
-		if sErr != nil {
-			err = sErr
-			t.Errorf("GetEntityModel failed. err:%s", err.Error())
-			return
-		}
-
-		sModelList[idx] = sModel
+	composeModel, composeErr = o1.Insert(composeModel)
+	if composeErr != nil {
+		err = composeErr
+		t.Errorf("Insert failed. err:%s", err.Error())
+		return
 	}
-	for idx := 0; idx < 100; idx++ {
-		vModel, vErr := o1.Update(sModelList[idx])
-		if vErr != nil {
-			err = vErr
-			t.Errorf("Update failed. err:%s", err.Error())
-			return
-		}
 
-		sObjectVal := vModel.Interface(true).(*remote.ObjectValue)
-		sVal := sValList[idx]
-		err = provider.UpdateEntity(sObjectVal, sVal)
-		if err != nil {
-			t.Errorf("UpdateEntity failed. err:%s", err.Error())
-			return
-		}
-		sValList[idx] = sVal
-		sModelList[idx] = vModel
-		sObjectValList[idx] = sObjectVal
+	composeObjectValue = composeModel.Interface(true).(*remote.ObjectValue)
+	err = helper.UpdateEntity(composeObjectValue, composePtr)
+	if err != nil {
+		t.Errorf("UpdateEntity failed. err:%s", err.Error())
+		return
+	}
+
+	composePtr.Name = "hi"
+	composeObjectValue, composeObjectErr = helper.GetObjectValue(composePtr)
+	if composeObjectErr != nil {
+		err = composeObjectErr
+		t.Errorf("GetObjectValue failed. err:%s", err.Error())
+		return
+	}
+
+	composeModel, composeErr = remoteProvider.GetEntityModel(composeObjectValue)
+	if composeErr != nil {
+		err = composeErr
+		t.Errorf("GetEntityModel failed. err:%s", err.Error())
+		return
+	}
+
+	vModel, vErr := o1.Update(composeModel)
+	if vErr != nil {
+		err = vErr
+		t.Errorf("Update failed. err:%s", err.Error())
+		return
+	}
+
+	composeObjectValue = vModel.Interface(true).(*remote.ObjectValue)
+	err = helper.UpdateEntity(composeObjectValue, composePtr)
+	if err != nil {
+		t.Errorf("UpdateEntity failed. err:%s", err.Error())
+		return
 	}
 
 	// query
-	qValList := []*Compose{}
-	qObjectValList := []*remote.ObjectValue{}
-	qModelList := []model.Model{}
-	for idx := 0; idx < 100; idx++ {
-		qVal := &Compose{
-			ID:             sValList[idx].ID,
-			PtrSimple:      &Simple{},
-			SimpleArray:    []Simple{},
-			SimplePtrArray: []*Simple{},
-			PtrSimpleArray: &[]Simple{},
-			PtrReference:   &Reference{},
-			RefArray:       []Reference{},
-			RefPtrArray:    []*Reference{},
-			PtrRefArray:    &[]*Reference{},
-			PtrCompose:     &Compose{},
-		}
-		qValList = append(qValList, qVal)
-
-		qObjectVal, qObjectErr := remote.GetObjectValue(qVal)
-		if qObjectErr != nil {
-			err = qObjectErr
-			t.Errorf("GetObjectValue failed. err:%s", err.Error())
-			return
-		}
-		qObjectValList = append(qObjectValList, qObjectVal)
-
-		qModel, qErr := remoteProvider.GetEntityModel(qObjectVal)
-		if qErr != nil {
-			err = qErr
-			t.Errorf("GetEntityModel failed. err:%s", err.Error())
-			return
-		}
-
-		qModelList = append(qModelList, qModel)
+	queryComposeVal := &Compose{
+		ID:                composePtr.ID,
+		SimplePtr:         &Simple{},
+		SimpleArray:       []Simple{},
+		SimplePtrArray:    []*Simple{},
+		SimpleArrayPtr:    &[]Simple{},
+		ReferencePtr:      &Reference{},
+		ReferenceArray:    []Reference{},
+		ReferencePtrArray: []*Reference{},
+		ComposePtr:        &Compose{},
 	}
 
-	for idx := 0; idx < 100; idx++ {
-		qModel, qErr := o1.Query(qModelList[idx])
-		if qErr != nil {
-			err = qErr
-			t.Errorf("Query failed. err:%s", err.Error())
-			return
-		}
-
-		qObjectVal := qModel.Interface(true).(*remote.ObjectValue)
-		qVal := qValList[idx]
-		err = provider.UpdateEntity(qObjectVal, qVal)
-		if err != nil {
-			t.Errorf("UpdateEntity failed. err:%s", err.Error())
-			return
-		}
-		qValList[idx] = qVal
-		qModelList[idx] = qModel
-		qObjectValList[idx] = qObjectVal
-	}
-
-	for idx := 0; idx < 100; idx++ {
-		sVal := sValList[idx]
-		qVal := qValList[idx]
-		if !sVal.IsSame(qVal) {
-			err = fmt.Errorf("compare value failed")
-			t.Errorf("IsSame failed. err:%s", err.Error())
-			return
-		}
-	}
-
-	bqValList := []*Compose{}
-	bqSliceObject, bqSliceErr := remote.GetSliceObjectValue(&bqValList)
-	if bqSliceErr != nil {
-		t.Errorf("GetSliceObjectValue failed, err:%s", bqSliceErr.Error())
-		return
-	}
-	bqModel, bqErr := remoteProvider.GetEntityModel(bqSliceObject)
-	if bqErr != nil {
-		t.Errorf("GetEntityModel failed, err:%s", bqErr.Error())
+	queryObjectValue, queryObjectErr := helper.GetObjectValue(queryComposeVal)
+	if queryObjectErr != nil {
+		err = queryObjectErr
+		t.Errorf("GetObjectValue failed. err:%s", err.Error())
 		return
 	}
 
-	filter := orm.GetFilter(remoteProvider)
-	filter.Equal("name", "hi")
-	filter.ValueMask(&Compose{
-		PtrSimple:      &Simple{},
-		SimpleArray:    []Simple{},
-		SimplePtrArray: []*Simple{},
-		PtrSimpleArray: &[]Simple{},
-		PtrReference:   &Reference{},
-		RefArray:       []Reference{},
-		RefPtrArray:    []*Reference{},
-		PtrRefArray:    &[]*Reference{},
-		PtrCompose:     &Compose{},
-	})
-	bqModelList, bqModelErr := o1.BatchQuery(bqModel, filter)
+	queryModel, queryErr := remoteProvider.GetEntityModel(queryObjectValue)
+	if queryErr != nil {
+		err = queryErr
+		t.Errorf("GetEntityModel failed. err:%s", err.Error())
+		return
+	}
+
+	queryModel, queryErr = o1.Query(queryModel)
+	if queryErr != nil {
+		err = queryErr
+		t.Errorf("Query failed. err:%s", err.Error())
+		return
+	}
+
+	queryObjectVal := queryModel.Interface(true).(*remote.ObjectValue)
+	err = helper.UpdateEntity(queryObjectVal, queryComposeVal)
+	if err != nil {
+		t.Errorf("UpdateEntity failed. err:%s", err.Error())
+		return
+	}
+
+	if !composePtr.IsSame(queryComposeVal) {
+		err = cd.NewError(cd.Unexpected, "compare value failed")
+		t.Errorf("IsSame failed. err:%s", err.Error())
+		return
+	}
+
+	composePtr = &Compose{}
+	composeObjectPtr, composeObjectErr := helper.GetObject(composePtr)
+	if composeObjectErr != nil {
+		t.Errorf("GetObject failed, err:%s", composeObjectErr.Error())
+		return
+	}
+
+	filter, err := remoteProvider.GetModelFilter(composeObjectPtr)
+	if err != nil {
+		t.Errorf("GetEntityFilter failed, err:%s", err.Error())
+		return
+	}
+
+	err = filter.Equal("name", "hi")
+	if err != nil {
+		t.Errorf("filter.Equal failed, err:%s", err.Error())
+		return
+	}
+
+	maskComposePtr := &Compose{
+		SimplePtr:         &Simple{},
+		SimpleArray:       []Simple{},
+		SimplePtrArray:    []*Simple{},
+		SimpleArrayPtr:    &[]Simple{},
+		ReferencePtr:      &Reference{},
+		ReferenceArray:    []Reference{},
+		ReferencePtrArray: []*Reference{},
+		ComposePtr:        &Compose{},
+	}
+
+	maskObjectValuePtr, maskObjectValueErr := helper.GetObjectValue(maskComposePtr)
+	if maskObjectValueErr != nil {
+		t.Errorf("GetObjectValue failed, err:%s", maskObjectValueErr.Error())
+		return
+	}
+
+	err = filter.ValueMask(maskObjectValuePtr)
+	if err != nil {
+		t.Errorf("filter.ValueMask failed, err:%s", err.Error())
+		return
+	}
+
+	bqModelList, bqModelErr := o1.BatchQuery(filter)
 	if bqModelErr != nil {
 		t.Errorf("BatchQuery failed, err:%s", bqModelErr.Error())
 		return
 	}
-	if len(bqModelList) != 100 {
+	if len(bqModelList) != 1 {
 		t.Errorf("batch query compose failed")
 		return
 	}
 
 	// delete
-	for idx := 0; idx < 100; idx++ {
-		_, qErr := o1.Delete(bqModelList[idx])
-		if qErr != nil {
-			err = qErr
-			return
-		}
+	_, qErr := o1.Delete(bqModelList[0])
+	if qErr != nil {
+		err = qErr
+		t.Errorf("o1.Delete failed, err:%s", qErr.Error())
+		return
 	}
 }
