@@ -3,26 +3,35 @@ package mysql
 import (
 	"fmt"
 
+	cd "github.com/muidea/magicCommon/def"
+	"github.com/muidea/magicCommon/foundation/log"
+
 	"github.com/muidea/magicOrm/model"
 )
 
-// BuildCount BuildCount
-func (s *Builder) BuildCount(filter model.Filter) (ret string, err error) {
-	pkField := s.modelInfo.GetPrimaryField()
-
-	ret = fmt.Sprintf("SELECT COUNT(%s) FROM `%s`", pkField.GetTag().GetName(), s.getHostTableName(s.modelInfo))
+// BuildCount build count
+func (s *Builder) BuildCount(vModel model.Model, filter model.Filter) (ret *ResultStack, err *cd.Error) {
+	resultStackPtr := &ResultStack{}
+	pkFieldName := vModel.GetPrimaryField().GetName()
+	countSQL := fmt.Sprintf("SELECT COUNT(`%s`) FROM `%s`", pkFieldName, s.buildCodec.ConstructModelTableName(vModel))
 	if filter != nil {
-		filterSQL, filterErr := s.buildFilter(filter)
+		filterSQL, filterErr := s.buildFilter(vModel, filter, resultStackPtr)
 		if filterErr != nil {
 			err = filterErr
+			log.Errorf("BuildCount failed, s.buildFilter error:%s", err.Error())
 			return
 		}
 
 		if filterSQL != "" {
-			ret = fmt.Sprintf("%s WHERE %s", ret, filterSQL)
+			countSQL = fmt.Sprintf("%s WHERE %s", countSQL, filterSQL)
 		}
 	}
 
-	//log.Print(ret)
+	if traceSQL() {
+		log.Infof("[SQL] count: %s", countSQL)
+	}
+
+	resultStackPtr.SetSQL(countSQL)
+	ret = resultStackPtr
 	return
 }
