@@ -2,39 +2,68 @@ package mysql
 
 import (
 	"fmt"
-	"log"
+
+	cd "github.com/muidea/magicCommon/def"
+	"github.com/muidea/magicCommon/foundation/log"
 
 	"github.com/muidea/magicOrm/model"
 )
 
 // BuildDelete  BuildDelete
-func (s *Builder) BuildDelete() (ret string, err error) {
-	filterStr, filterErr := s.buildPKFilter(s.modelInfo)
+func (s *Builder) BuildDelete(vModel model.Model) (ret *ResultStack, err *cd.Error) {
+	resultStackPtr := &ResultStack{}
+	filterStr, filterErr := s.buildFieldFilter(vModel.GetPrimaryField(), resultStackPtr)
 	if filterErr != nil {
 		err = filterErr
-		log.Printf("buildPKFilter failed, err:%s", err.Error())
+		log.Errorf("BuildDelete failed, s.BuildModelFilter error:%s", err.Error())
 		return
 	}
 
-	ret = fmt.Sprintf("DELETE FROM `%s` WHERE %s", s.getHostTableName(s.modelInfo), filterStr)
-	//log.Print(ret)
+	deleteSQL := fmt.Sprintf("DELETE FROM `%s` WHERE %s", s.buildCodec.ConstructModelTableName(vModel), filterStr)
+	if traceSQL() {
+		log.Infof("[SQL] delete: %s", deleteSQL)
+	}
 
+	resultStackPtr.SetSQL(deleteSQL)
+	ret = resultStackPtr
 	return
 }
 
 // BuildDeleteRelation BuildDeleteRelation
-func (s *Builder) BuildDeleteRelation(fieldName string, relationInfo model.Model) (delRight, delRelation string, err error) {
-	leftVal, leftErr := s.getModelValue(s.modelInfo)
-	if leftErr != nil {
-		err = leftErr
+func (s *Builder) BuildDeleteRelation(vModel model.Model, vField model.Field) (delHost, delRelation *ResultStack, err *cd.Error) {
+	hostVal := vModel.GetPrimaryField().GetValue().Get()
+	relationTableName, relationErr := s.buildCodec.ConstructRelationTableName(vModel, vField)
+	if relationErr != nil {
+		err = relationErr
+		log.Errorf("BuildDeleteRelation %s failed, s.buildCodec.ConstructRelationTableName error:%s", vField.GetName(), err.Error())
 		return
 	}
 
-	delRight = fmt.Sprintf("DELETE FROM `%s` WHERE `id` in (SELECT `right` FROM `%s` WHERE `left`=%v)", s.getHostTableName(relationInfo), s.GetRelationTableName(fieldName, relationInfo), leftVal)
-	//log.Print(delRight)
+	rModel, rErr := s.modelProvider.GetTypeModel(vField.GetType())
+	if rErr != nil {
+		err = rErr
+		log.Errorf("BuildDeleteRelation %s failed, s.modelProvider.GetTypeModel error:%s", vField.GetName(), err.Error())
+		return
+	}
 
-	delRelation = fmt.Sprintf("DELETE FROM `%s` WHERE `left`=%v", s.GetRelationTableName(fieldName, relationInfo), leftVal)
-	//log.Print(delRelation)
+	delHostStackPtr := &ResultStack{}
+	delHostSQL := fmt.Sprintf("DELETE FROM `%s` WHERE `%s` IN (SELECT `right` FROM `%s` WHERE `left`=?)",
+		s.buildCodec.ConstructModelTableName(vField.GetType()),
+		rModel.GetPrimaryField().GetName(),
+		relationTableName)
+	delHostStackPtr.SetSQL(delHostSQL)
+	delHostStackPtr.PushArgs(hostVal)
+	delHost = delHostStackPtr
+
+	delRelationStackPtr := &ResultStack{}
+	delRelationSQL := fmt.Sprintf("DELETE FROM `%s` WHERE `left`=?", relationTableName)
+	delRelationStackPtr.SetSQL(delRelationSQL)
+	delRelationStackPtr.PushArgs(hostVal)
+	delRelation = delRelationStackPtr
+
+	if traceSQL() {
+		log.Infof("[SQL] delete host: %s, delete relation: %s", delHostSQL, delRelationSQL)
+	}
 
 	return
 }
