@@ -3,98 +3,79 @@ package mysql
 import (
 	"fmt"
 
+	cd "github.com/muidea/magicCommon/def"
+	"github.com/muidea/magicCommon/foundation/log"
+
 	"github.com/muidea/magicOrm/model"
 )
 
-// BuildInsert  BuildInsert
-func (s *Builder) BuildInsert() (ret string, err error) {
-	sql := ""
-	fieldNames, fieldErr := s.getFieldInsertNames(s.modelInfo)
-	if fieldErr != nil {
-		err = fieldErr
-		return
-	}
-
-	fieldValues, fieldErr := s.getFieldInsertValues(s.modelInfo)
-	if fieldErr != nil {
-		err = fieldErr
-		return
-	}
-
-	sql = fmt.Sprintf("INSERT INTO `%s` (%s) VALUES (%s)", s.getHostTableName(s.modelInfo), fieldNames, fieldValues)
-	//log.Print(sql)
-	ret = sql
-
-	return
-}
-
-// BuildInsertRelation BuildInsertRelation
-func (s *Builder) BuildInsertRelation(fieldName string, relationInfo model.Model) (ret string, err error) {
-	leftVal, rightVal, valErr := s.getRelationValue(relationInfo)
-	if valErr != nil {
-		err = valErr
-		return
-	}
-
-	ret = fmt.Sprintf("INSERT INTO `%s` (`left`, `right`) VALUES (%v,%v);", s.GetRelationTableName(fieldName, relationInfo), leftVal, rightVal)
-	//log.Print(ret)
-
-	return
-}
-
-func (s *Builder) getFieldInsertNames(info model.Model) (ret string, err error) {
-	str := ""
-	for _, field := range info.GetFields() {
-		fTag := field.GetTag()
-		if fTag.IsAutoIncrement() {
+// BuildInsert  Build Insert
+func (s *Builder) BuildInsert(vModel model.Model) (ret *ResultStack, err *cd.Error) {
+	resultStackPtr := &ResultStack{}
+	fieldNames := ""
+	fieldValues := ""
+	for _, field := range vModel.GetFields() {
+		if !model.IsBasicField(field) || !model.IsValidField(field) {
 			continue
 		}
 
-		fType := field.GetType()
+		fSpec := field.GetSpec()
+		if fSpec.GetValueDeclare() == model.AutoIncrement {
+			continue
+		}
+
 		fValue := field.GetValue()
-		if !fType.IsBasic() || fValue.IsNil() {
-			continue
-		}
-
-		if str == "" {
-			str = fmt.Sprintf("`%s`", fTag.GetName())
-		} else {
-			str = fmt.Sprintf("%s,`%s`", str, fTag.GetName())
-		}
-	}
-
-	ret = str
-	return
-}
-
-func (s *Builder) getFieldInsertValues(info model.Model) (ret string, err error) {
-	str := ""
-	for _, field := range info.GetFields() {
-		fTag := field.GetTag()
-		if fTag.IsAutoIncrement() {
-			continue
-		}
-
-		fType := field.GetType()
-		fValue := field.GetValue()
-		if !fType.IsBasic() || fValue.IsNil() {
-			continue
-		}
-
-		fStr, fErr := s.buildValue(fValue, fType)
-		if fErr != nil {
-			err = fErr
+		encodeVal, encodeErr := s.buildCodec.PackedBasicFieldValue(field, fValue)
+		if encodeErr != nil {
+			err = encodeErr
+			log.Errorf("BuildInsert %s failed, encodeFieldValue error:%s", field.GetName(), err.Error())
 			return
 		}
 
-		if str == "" {
-			str = fmt.Sprintf("%v", fStr)
+		if fieldNames == "" {
+			fieldNames = fmt.Sprintf("`%s`", field.GetName())
 		} else {
-			str = fmt.Sprintf("%s,%v", str, fStr)
+			fieldNames = fmt.Sprintf("%s,`%s`", fieldNames, field.GetName())
+		}
+
+		resultStackPtr.PushArgs(encodeVal)
+		if fieldValues == "" {
+			fieldValues = fmt.Sprintf("%v", "?")
+		} else {
+			fieldValues = fmt.Sprintf("%s,%v", fieldValues, "?")
 		}
 	}
 
-	ret = str
+	insertSQL := fmt.Sprintf("INSERT INTO `%s` (%s) VALUES (%s)", s.buildCodec.ConstructModelTableName(vModel), fieldNames, fieldValues)
+	if traceSQL() {
+		log.Infof("[SQL] insert: %s", insertSQL)
+	}
 
+	resultStackPtr.SetSQL(insertSQL)
+	ret = resultStackPtr
+	return
+}
+
+// BuildInsertRelation Build Insert Relation
+func (s *Builder) BuildInsertRelation(vModel model.Model, vField model.Field, rModel model.Model) (ret *ResultStack, err *cd.Error) {
+	relationTableName, relationErr := s.buildCodec.ConstructRelationTableName(vModel, vField)
+	if relationErr != nil {
+		err = relationErr
+		log.Errorf("BuildInsertRelation %s failed, s.buildCodec.ConstructRelationTableName error:%s", vField.GetName(), err.Error())
+		return
+	}
+
+	leftVal := vModel.GetPrimaryField().GetValue().Get()
+	rightVal := rModel.GetPrimaryField().GetValue().Get()
+	resultStackPtr := &ResultStack{}
+	insertRelationSQL := fmt.Sprintf("INSERT INTO `%s` (`left`, `right`) VALUES (?,?)", relationTableName)
+	resultStackPtr.PushArgs(leftVal, rightVal)
+	resultStackPtr.SetSQL(insertRelationSQL)
+	//log.Print(ret)
+	if traceSQL() {
+		log.Infof("[SQL] insert relation: %s", insertRelationSQL)
+	}
+
+	ret = resultStackPtr
 	return
 }
