@@ -4,428 +4,171 @@ import (
 	"fmt"
 	"reflect"
 
+	cd "github.com/muidea/magicCommon/def"
+	"github.com/muidea/magicCommon/foundation/log"
 	"github.com/muidea/magicOrm/model"
-	"github.com/muidea/magicOrm/util"
 )
 
 var _declareObjectSliceValue SliceObjectValue
 var _declareObjectValue ObjectValue
 
-func getSliceType(tType model.Type) (ret reflect.Type, err error) {
+func getSliceInitValue(tType model.Type) (ret any) {
 	eType := tType.Elem()
 	switch eType.GetValue() {
-	case util.TypeBooleanField:
-		if tType.IsPtrType() {
-			if eType.IsPtrType() {
-				var val *[]*bool
-				ret = reflect.TypeOf(val)
-				return
-			}
-			var val *[]bool
-			ret = reflect.TypeOf(val)
-			return
-		}
-		if eType.IsPtrType() {
-			var val []*bool
-			ret = reflect.TypeOf(val)
-			return
-		}
-		var val []bool
-		ret = reflect.TypeOf(val)
-	case util.TypeStringField,
-		util.TypeDateTimeField:
-		if tType.IsPtrType() {
-			if eType.IsPtrType() {
-				var val *[]*string
-				ret = reflect.TypeOf(val)
-				return
-			}
-			var val *[]string
-			ret = reflect.TypeOf(val)
-			return
-		}
-		if eType.IsPtrType() {
-			var val []*string
-			ret = reflect.TypeOf(val)
-			return
-		}
-		var val []string
-		ret = reflect.TypeOf(val)
-	case util.TypeBitField:
-		if tType.IsPtrType() {
-			if eType.IsPtrType() {
-				var val *[]*int8
-				ret = reflect.TypeOf(val)
-				return
-			}
-			var val *[]int8
-			ret = reflect.TypeOf(val)
-			return
-		}
-		if eType.IsPtrType() {
-			var val []*int8
-			ret = reflect.TypeOf(val)
-			return
-		}
-		var val []int8
-		ret = reflect.TypeOf(val)
-	case util.TypeSmallIntegerField:
-		if tType.IsPtrType() {
-			if eType.IsPtrType() {
-				var val *[]*int16
-				ret = reflect.TypeOf(val)
-				return
-			}
-			var val *[]int16
-			ret = reflect.TypeOf(val)
-			return
-		}
-		if eType.IsPtrType() {
-			var val []*int16
-			ret = reflect.TypeOf(val)
-			return
-		}
-		var val []int16
-		ret = reflect.TypeOf(val)
-	case util.TypeInteger32Field:
-		if tType.IsPtrType() {
-			if eType.IsPtrType() {
-				var val *[]*int32
-				ret = reflect.TypeOf(val)
-				return
-			}
-			var val *[]int32
-			ret = reflect.TypeOf(val)
-			return
-		}
-		if eType.IsPtrType() {
-			var val []*int32
-			ret = reflect.TypeOf(val)
-			return
-		}
-		var val []int32
-		ret = reflect.TypeOf(val)
-	case util.TypeIntegerField:
-		if tType.IsPtrType() {
-			if eType.IsPtrType() {
-				var val *[]*int
-				ret = reflect.TypeOf(val)
-				return
-			}
-			var val *[]int
-			ret = reflect.TypeOf(val)
-			return
-		}
-		if eType.IsPtrType() {
-			var val []*int
-			ret = reflect.TypeOf(val)
-			return
-		}
-		var val []int
-		ret = reflect.TypeOf(val)
-	case util.TypeBigIntegerField:
-		if tType.IsPtrType() {
-			if eType.IsPtrType() {
-				var val *[]*int64
-				ret = reflect.TypeOf(val)
-				return
-			}
-			var val *[]int64
-			ret = reflect.TypeOf(val)
-			return
-		}
-		if eType.IsPtrType() {
-			var val []*int64
-			ret = reflect.TypeOf(val)
-			return
-		}
-		var val []int64
-		ret = reflect.TypeOf(val)
-	case util.TypePositiveBitField:
-		if tType.IsPtrType() {
-			if eType.IsPtrType() {
-				var val *[]*uint8
-				ret = reflect.TypeOf(val)
-				return
-			}
-			var val *[]uint8
-			ret = reflect.TypeOf(val)
-			return
-		}
-		if eType.IsPtrType() {
-			var val []*uint8
-			ret = reflect.TypeOf(val)
-			return
-		}
-		var val []uint8
-		ret = reflect.TypeOf(val)
-	case util.TypePositiveSmallIntegerField:
-		if tType.IsPtrType() {
-			if eType.IsPtrType() {
-				var val *[]*uint16
-				ret = reflect.TypeOf(val)
-				return
-			}
-			var val *[]uint16
-			ret = reflect.TypeOf(val)
-			return
-		}
-		if eType.IsPtrType() {
-			var val []*uint16
-			ret = reflect.TypeOf(val)
-			return
-		}
-		var val []uint16
-		ret = reflect.TypeOf(val)
-	case util.TypePositiveInteger32Field:
-		if tType.IsPtrType() {
-			if eType.IsPtrType() {
-				var val *[]*uint32
-				ret = reflect.TypeOf(val)
-				return
-			}
-			var val *[]uint32
-			ret = reflect.TypeOf(val)
-			return
-		}
-		if eType.IsPtrType() {
-			var val []*uint32
-			ret = reflect.TypeOf(val)
-			return
-		}
-		var val []uint32
-		ret = reflect.TypeOf(val)
-	case util.TypePositiveIntegerField:
-		if tType.IsPtrType() {
-			if eType.IsPtrType() {
-				var val *[]*uint
-				ret = reflect.TypeOf(val)
-				return
-			}
-			var val *[]uint
-			ret = reflect.TypeOf(val)
-			return
-		}
-		if eType.IsPtrType() {
-			var val []*uint
-			ret = reflect.TypeOf(val)
-			return
-		}
-		var val []uint
-		ret = reflect.TypeOf(val)
-	case util.TypePositiveBigIntegerField:
-		if tType.IsPtrType() {
-			if eType.IsPtrType() {
-				var val *[]*uint64
-				ret = reflect.TypeOf(val)
-				return
-			}
-			var val *[]uint64
-			ret = reflect.TypeOf(val)
-			return
-		}
-		if eType.IsPtrType() {
-			var val []*uint64
-			ret = reflect.TypeOf(val)
-			return
-		}
-		var val []uint64
-		ret = reflect.TypeOf(val)
-	case util.TypeFloatField:
-		if tType.IsPtrType() {
-			if eType.IsPtrType() {
-				var val *[]*float32
-				ret = reflect.TypeOf(val)
-				return
-			}
-			var val *[]float32
-			ret = reflect.TypeOf(val)
-			return
-		}
-		if eType.IsPtrType() {
-			var val []*float32
-			ret = reflect.TypeOf(val)
-			return
-		}
-		var val []float32
-		ret = reflect.TypeOf(val)
-	case util.TypeDoubleField:
-		if tType.IsPtrType() {
-			if eType.IsPtrType() {
-				var val *[]*float64
-				ret = reflect.TypeOf(val)
-				return
-			}
-			var val *[]float64
-			ret = reflect.TypeOf(val)
-			return
-		}
-		if eType.IsPtrType() {
-			var val []*float64
-			ret = reflect.TypeOf(val)
-			return
-		}
-		var val []float64
-		ret = reflect.TypeOf(val)
-	case util.TypeStructField:
-		ret = reflect.TypeOf(&_declareObjectSliceValue)
+	case model.TypeBooleanValue:
+		ret = []bool{}
+	case model.TypeStringValue,
+		model.TypeDateTimeValue:
+		ret = []string{}
+	case model.TypeBitValue:
+		ret = []int8{}
+	case model.TypeSmallIntegerValue:
+		ret = []int16{}
+	case model.TypeInteger32Value:
+		ret = []int32{}
+	case model.TypeIntegerValue:
+		ret = []int{}
+	case model.TypeBigIntegerValue:
+		ret = []int64{}
+	case model.TypePositiveBitValue:
+		ret = []uint8{}
+	case model.TypePositiveSmallIntegerValue:
+		ret = []uint16{}
+	case model.TypePositiveInteger32Value:
+		ret = []uint32{}
+	case model.TypePositiveIntegerValue:
+		ret = []uint{}
+	case model.TypePositiveBigIntegerValue:
+		ret = []uint64{}
+	case model.TypeFloatValue:
+		ret = []float32{}
+	case model.TypeDoubleValue:
+		ret = []float64{}
 	default:
-		err = fmt.Errorf("unexpect slice item type, name:%s, type:%d", tType.GetName(), tType.GetValue())
+		err := fmt.Errorf("unexpected slice item type, name:%s, type:%d", tType.GetName(), tType.GetValue())
+		panic(err)
 	}
 
 	return
 }
 
-func getType(tType model.Type) (ret reflect.Type, err error) {
+func getBasicInitValue(tType model.Type) (ret any) {
 	switch tType.GetValue() {
-	case util.TypeBooleanField:
-		if tType.IsPtrType() {
-			var val *bool
-			ret = reflect.TypeOf(val)
-			return
-		}
-		var val bool
-		ret = reflect.TypeOf(val)
-	case util.TypeStringField,
-		util.TypeDateTimeField:
-		if tType.IsPtrType() {
-			var val *string
-			ret = reflect.TypeOf(val)
-			return
-		}
-		var val string
-		ret = reflect.TypeOf(val)
-	case util.TypeBitField:
-		if tType.IsPtrType() {
-			var val *int8
-			ret = reflect.TypeOf(val)
-			return
-		}
-		var val int8
-		ret = reflect.TypeOf(val)
-	case util.TypeSmallIntegerField:
-		if tType.IsPtrType() {
-			var val *int16
-			ret = reflect.TypeOf(val)
-			return
-		}
-		var val int16
-		ret = reflect.TypeOf(val)
-	case util.TypeInteger32Field:
-		if tType.IsPtrType() {
-			var val *int32
-			ret = reflect.TypeOf(val)
-			return
-		}
-		var val int32
-		ret = reflect.TypeOf(val)
-	case util.TypeIntegerField:
-		if tType.IsPtrType() {
-			var val *int
-			ret = reflect.TypeOf(val)
-			return
-		}
-		var val int
-		ret = reflect.TypeOf(val)
-	case util.TypeBigIntegerField:
-		if tType.IsPtrType() {
-			var val *int64
-			ret = reflect.TypeOf(val)
-			return
-		}
-		var val int64
-		ret = reflect.TypeOf(val)
-	case util.TypePositiveBitField:
-		if tType.IsPtrType() {
-			var val *uint8
-			ret = reflect.TypeOf(val)
-			return
-		}
-		var val uint8
-		ret = reflect.TypeOf(val)
-	case util.TypePositiveSmallIntegerField:
-		if tType.IsPtrType() {
-			var val *uint16
-			ret = reflect.TypeOf(val)
-			return
-		}
-		var val uint16
-		ret = reflect.TypeOf(val)
-	case util.TypePositiveInteger32Field:
-		if tType.IsPtrType() {
-			var val *uint32
-			ret = reflect.TypeOf(val)
-			return
-		}
-		var val uint32
-		ret = reflect.TypeOf(val)
-	case util.TypePositiveIntegerField:
-		if tType.IsPtrType() {
-			var val *uint
-			ret = reflect.TypeOf(val)
-			return
-		}
-		var val uint
-		ret = reflect.TypeOf(val)
-	case util.TypePositiveBigIntegerField:
-		if tType.IsPtrType() {
-			var val *uint64
-			ret = reflect.TypeOf(val)
-			return
-		}
-		var val uint64
-		ret = reflect.TypeOf(val)
-	case util.TypeFloatField:
-		if tType.IsPtrType() {
-			var val *float32
-			ret = reflect.TypeOf(val)
-			return
-		}
-		var val float32
-		ret = reflect.TypeOf(val)
-	case util.TypeDoubleField:
-		if tType.IsPtrType() {
-			var val *float64
-			ret = reflect.TypeOf(val)
-		}
-		var val float64
-		ret = reflect.TypeOf(val)
-	case util.TypeStructField:
-		ret = reflect.TypeOf(&_declareObjectValue)
-	case util.TypeSliceField:
-		ret, err = getSliceType(tType)
+	case model.TypeBooleanValue:
+		ret = false
+	case model.TypeStringValue,
+		model.TypeDateTimeValue:
+		ret = ""
+	case model.TypeBitValue:
+		ret = int8(0)
+	case model.TypeSmallIntegerValue:
+		ret = int16(0)
+	case model.TypeInteger32Value:
+		ret = int32(0)
+	case model.TypeIntegerValue:
+		ret = 0
+	case model.TypeBigIntegerValue:
+		ret = int64(0)
+	case model.TypePositiveBitValue:
+		ret = uint8(0)
+	case model.TypePositiveSmallIntegerValue:
+		ret = uint16(0)
+	case model.TypePositiveInteger32Value:
+		ret = uint32(0)
+	case model.TypePositiveIntegerValue:
+		ret = uint(0)
+	case model.TypePositiveBigIntegerValue:
+		ret = uint64(0)
+	case model.TypeFloatValue:
+		ret = float32(0.00)
+	case model.TypeDoubleValue:
+		ret = 0.00
+	case model.TypeSliceValue:
+		ret = getSliceInitValue(tType)
 	default:
-		err = fmt.Errorf("unexpect item type, name:%s, type:%d", tType.GetName(), tType.GetValue())
+		err := fmt.Errorf("unexpected basic item type, name:%s, type:%d", tType.GetName(), tType.GetValue())
+		panic(err)
 	}
 
 	return
 }
 
-func getInitializeValue(tType model.Type) (ret reflect.Value, err error) {
-	cType, cErr := getType(tType)
-	if cErr != nil {
-		err = cErr
+func getStructInitValue(tType model.Type) (ret any) {
+	if model.IsSliceType(tType.GetValue()) {
+		sliceVal := _declareObjectSliceValue.Copy()
+		sliceVal.Name = tType.GetName()
+		sliceVal.PkgPath = tType.GetPkgPath()
+		ret = sliceVal
 		return
 	}
 
-	if tType.IsPtrType() || !tType.IsBasic() {
-		cType = cType.Elem()
+	if model.IsStructType(tType.GetValue()) {
+		valPtr := _declareObjectValue.Copy()
+		valPtr.Name = tType.GetName()
+		valPtr.PkgPath = tType.GetPkgPath()
+		ret = valPtr
+		return
 	}
 
-	cValue := reflect.New(cType).Elem()
-	if !tType.IsBasic() {
-		cValue.FieldByName("Name").SetString(tType.GetName())
-		cValue.FieldByName("PkgPath").SetString(tType.GetPkgPath())
-		cValue.FieldByName("IsPtr").SetBool(tType.IsPtrType())
-		if util.IsSliceType(tType.GetValue()) {
-			cValue.FieldByName("IsElemPtr").SetBool(tType.Elem().IsPtrType())
-		}
+	return
+}
+
+func getInitializeValue(tType model.Type) (ret any) {
+	if !model.IsBasic(tType) {
+		ret = getStructInitValue(tType)
+		return
 	}
 
-	if tType.IsPtrType() || !tType.IsBasic() {
-		cValue = cValue.Addr()
+	ret = getBasicInitValue(tType)
+	return
+}
+
+func rewriteObjectValue(rawPtr *ObjectValue, newPtr *ObjectValue) (err *cd.Error) {
+	if rawPtr == nil || newPtr == nil {
+		return
+	}
+	if rawPtr.PkgPath != newPtr.PkgPath {
+		err = cd.NewError(cd.Unexpected, "illegal object value")
+		return
 	}
 
-	ret = cValue
+	rawPtr.Fields = newPtr.Fields
+	return
+}
+
+func rewriteSliceObjectValue(rawPtr *SliceObjectValue, newPtr *SliceObjectValue) (err *cd.Error) {
+	if rawPtr == nil || newPtr == nil {
+		return
+	}
+	if rawPtr.PkgPath != newPtr.PkgPath {
+		err = cd.NewError(cd.Unexpected, "illegal slice object value")
+		return
+	}
+
+	rawPtr.Values = newPtr.Values
+	return
+}
+
+func appendBasicValue(sliceVal, val any) (ret any, err *cd.Error) {
+	rVal := reflect.ValueOf(sliceVal)
+	if rVal.Kind() != reflect.Slice {
+		err = cd.NewError(cd.Unexpected, "value is not slice")
+		log.Warnf("Append failed, value is not slice")
+		return
+	}
+
+	// Check if the types are compatible
+	elemType := rVal.Type().Elem()
+	valType := reflect.TypeOf(val)
+
+	if !valType.AssignableTo(elemType) {
+		err = cd.NewError(cd.Unexpected, "type mismatch, expected: "+elemType.String()+", got: "+valType.String())
+		log.Warnf("Append failed, type mismatch, expected: %s, got: %s", elemType, valType)
+		return
+	}
+
+	// Create a new slice with the appended value
+	newSlice := reflect.Append(rVal, reflect.ValueOf(val))
+	ret = newSlice.Interface()
 	return
 }
